% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkern_computation.R
\name{pkern_agg}
\alias{pkern_agg}
\title{Reduce the resolution of a regular grid by separable convolution}
\usage{
pkern_agg(z, dims = NULL, afact = 2, pars = NULL, tol = 0.05)
}
\arguments{
\item{z}{either a numeric matrix with dimensions \code{dims}, or its column-vectorization}

\item{dims}{integer vector c(nx, ny), the grid size}

\item{afact}{integer or vector of two integers, the aggregation factors}

\item{pars}{(optional) list of "x" and "y" kernel parameters}

\item{tol}{numeric between 0 and 1, tolerance for NAs}
}
\value{
numeric matrix of dimensions \code{floor(dims/afact)}, the aggregated grid
}
\description{
Aggregates the data in \code{z} by convolving it with the separable spatial kernel
defined by \code{pars}, producing a smaller output matrix containing every \code{afact}th
grid line. ie. only the x grid lines numbered 1, afact\link{1}, 2*afact\link{1}, ... etc.
}
\details{
This is a heuristic algorithm intended for reducing the size of images before
sending them to \code{graphics::image}.
}
