% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkern_raster.R
\name{pkern_checkRaster}
\alias{pkern_checkRaster}
\title{Check if terra or raster package is loaded and print error message otherwise}
\usage{
pkern_checkRaster()
}
\value{
logical vector, indicating if the package is loaded
}
\description{
Check if terra or raster package is loaded and print error message otherwise
}
\examples{
pkern_checkRaster()
}
