% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkern_index.R
\name{pkern_coords}
\alias{pkern_coords}
\title{Return coordinates of a grid of points in column-vectorized order}
\usage{
pkern_coords(g, out = "matrix", corner = FALSE)
}
\arguments{
\item{g}{any other object accepted/returned by \code{pkern_grid}}

\item{out}{character indicating return value type, either 'list', 'matrix', or 'sf'}

\item{corners}{logical, indicates to only return the corner points}
}
\value{
a matrix, list, or sf POINT collection, in column vectorized order
}
\description{
Expands sets of y and x grid line locations in the column-major order i that
pkern uses for grid data. This ordering sorts the data into blocks of length equal
to the vertical dimension of the grid, where: y descends within blocks within blocks
and x increases among blocks.
}
\details{
This is similar to \code{base::expand.grid} but with the first dimension descending instead
of ascending. Optionally, setting \code{out='sf'} builds an \code{sf} simple features object
from the coordinates and populates it with data (if any) from \code{g}. This can be slow for
large grids.
}
\examples{

# 5x5 grid example
str( pkern_coords(5) )
pkern_coords(5, out='list')
pkern_coords(5, corner=TRUE)
pkern_coords(5, corner=TRUE, out='list')

# 3x2 example and sf output type
pkern_coords(c(3,2))
if( requireNamespace('sf') ) {
pkern_coords(c(13, 12), out='sf')
}

}
