% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkern_covariance.R
\name{pkern_corr}
\alias{pkern_corr}
\title{Stationary 1D correlation kernels}
\usage{
pkern_corr(d, pars)
}
\arguments{
\item{d}{numeric vector of length n, the distances to evaluate}

\item{pars}{list with elements "k", the kernel name, and "p" the parameter vector}
}
\value{
the length-n vector of correlations
}
\description{
Computes stationary correlation function values for (nonegative) 1-dimensional
distances \code{d}. Parameter list entry \code{pars$kp} supplies the kernel parameter(s),
where \code{pars$k} is one of the following kernel names
}
\details{
"exp": exponential (special case of "gex" with p=1 and \code{d<-abs(d)}). \code{kp = rho}
"gau": gaussian/stable (special case of "gex" with p=2). \code{kp = rho}
"sph": spherical (AKA stable/Gaussian for p=2). \code{kp = rho}
"gex": gamma-exponential. \code{kp = c(rho, p)}
"mat": Whittle-Matern ((Handcock and Wallis parameterization). \code{kp = c(rho, kap)}
}
\examples{
pkern_corr(d=1:10, pars=list(k='exp', kp=5))
pkern_corr(d=1:10, pars=list(k='mat', kp=c(5, 3)))
}
