% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkern_var.R
\name{pkern_corr_mat}
\alias{pkern_corr_mat}
\title{Construct 1D stationary correlation matrices for regularly spaced data}
\usage{
pkern_corr_mat(pars, n, gres = 1, i = seq(n), j = seq(n))
}
\arguments{
\item{pars}{list of kernel parameters "k" and "kp" (see \code{pkern_corr})}

\item{n}{positive integer, the number of points on the 1D line}

\item{gres}{positive numeric, the distance between adjacent grid lines}

\item{i}{vector, a subset of \code{seq(n)} indicating rows to return}

\item{j}{vector, a subset of \code{seq(n)} indicating columns to return}
}
\value{
the n x n correlation matrix, or its subset as specified in \code{i}, \code{j}
}
\description{
An effient implementation that uses symmetry and Toeplitz structure arising
from assumption of stationarity of the random field and regularity of the grid.
}
\details{
\code{gres} scales the distance between adjacent points
}
\examples{
pars_x = pkern_pars(10, 'gxp')[['x']]
pkern_corr_mat(pars_x, n=10)
pkern_corr_mat(pars_x, n=10, i=2:4, j=2:4)
pkern_corr_mat(pars_x, n=3)
pkern_corr_mat(pars_x, n=3, gres=2)
}
