% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkern_covariance.R
\name{pkern_corrmat}
\alias{pkern_corrmat}
\title{Construct 1D stationary correlation matrices for regularly spaced data}
\usage{
pkern_corrmat(pars, n, ds = 1, i = seq(n), j = seq(n))
}
\arguments{
\item{pars}{list or parameters (in the form accepted by \code{pkern_corr})}

\item{n}{positive integer, the number of points on the 1D line}

\item{ds}{positive numeric, the distance between adjacent grid lines}

\item{i}{vector, a subset of \code{seq(n)} indicating rows to return}

\item{j}{vector, a subset of \code{seq(n)} indicating columns to return}
}
\value{
the n x n correlation matrix, or its subset as specified in \code{i}, \code{j}
}
\description{
An effient implementation that uses symmetry and Toeplitz structure arising
from assumption of stationarity.
}
\examples{
pars = pkern_corr('exp')
pkern_corrmat(pars, n=10)
pkern_corrmat(pars, n=10, i=2:4, j=2:4)
pkern_corrmat(pars, n=3)
pkern_corrmat(pars, n=3, ds=2)
}
