% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkern_snap.R
\name{pkern_estimate_sep}
\alias{pkern_estimate_sep}
\title{Estimate separation distance between grid lines in 2-dimensional grids}
\usage{
pkern_estimate_sep(
  g,
  pts,
  nmax = 100,
  local_search = TRUE,
  penalty = 1,
  control = NULL
)
}
\value{
the estimated value of \code{sep}
}
\description{
Estimates the subgrid resolution best matching an irregular set of points.
Helper function for \code{pkern_snap_1d}
TODO: flesh out documentation here and tidy inner loop in code
}
\examples{
sep = 25
ng = 10*sep
g = seq(ng)
sg = seq(sample(sep, 1), ng, by=sep)

# sample part of the subgrid and add noise
npts = round( (3/4) * length(sg) )
pts = sort(sample(sg, npts)) + stats::rnorm(npts)
pkern_snap_plot(g, pts)

# estimate separation and plot result
sep = pkern_estimate_sep(g, pts)
sep
snap = pkern_snap_1d(g, pts, sep)
pkern_snap_plot(snap, pts)

# add duplication
ndupe = 10
pts = sapply(pts, \(p) p + stats::rnorm(ndupe)) |> c()

# fit again
pkern_snap_plot(g, pts)
sep = pkern_estimate_sep(g, pts, distinct=FALSE)
sep
snap = pkern_snap_1d(g, pts, sep=sep, distinct=FALSE)
pkern_snap_plot(snap, pts)

}
