% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkern_grid.R
\name{pkern_export}
\alias{pkern_export}
\title{Convert column-vectorized grid to SpatRaster}
\usage{
pkern_export(g, template = "terra")
}
\arguments{
\item{g}{any object accepted or returned by \code{pkern_grid}}

\item{template}{character or RasterLayer/SpatRaster to set output type

Converts a column-vectorized vector or matrix to a SpatRaster, or if terra is
unavailable, a RasterLayer.}
}
\value{
a RasterLayer or SpatRaster containing the data from \code{g} (or a sub-grid)
}
\description{
Convert column-vectorized grid to SpatRaster
}
\examples{

if( requireNamespace('raster') ) {

# open example file as RasterLayer
r_path = system.file('external/rlogo.grd', package='raster')
r = raster::raster(r_path, band=1)
g = pkern_grid(r)

# convert back to RasterLayer and compare
r_from_g = pkern_export(g, 'raster')
print(r_from_g)
print(r)

# layer name, band number, and various other metadata are lost
all.equal(r_from_g, r)

# same with terra
if( requireNamespace('terra') ) {

# SpatRaster is the default return class (when terra loaded)
r = terra::rast(r_path)
g = pkern_grid(r)
r_from_g = pkern_export(g)

# notice only the first band is loaded by pkern_grid
print(r_from_g)
print(r)
}
}

}
