% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkern_estim.R
\name{pkern_fit}
\alias{pkern_fit}
\title{Fit a covariance model to the data by maximum likelihood}
\usage{
pkern_fit(g_obs, pars = NULL, X = NA, iso = TRUE, initial = NULL)
}
\arguments{
\item{X}{}
}
\value{

}
\description{
An automated model fitting procedure
}
\examples{

# define a grid
gdim = c(50, 53)
g_empty = pkern_grid(gdim)
pars_src = pkern_pars(g_empty)
pars_src = modifyList(pars_src, list(eps=runif(1, 0, 1e1), psill=runif(1, 0, 1e2)))
pars_src[['y']][['kp']] = pars_src[['x']][['kp']] = runif(1, 1, 50)

# generate example data and fit to it
gval = pkern_sim(g_empty, pars_src, quiet=TRUE)
g_obs = modifyList(g_empty, list(gval=gval))
pkern_plot(g_obs)
fit_result = pkern_fit(g_obs, pars='mat')

fit_result$pars |> pkern_pars_update()
pars_src |> pkern_pars_update()

pars2 = fit_result$pars
pars2$psill = NA
fit_result2 = pkern_fit(g_obs, pars=pars2)
fit_result2$pars |> pkern_pars_update()
pars_src |> pkern_pars_update()


is_fitted = is.na(pkern_pars_update(pars2))



}
