% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkern_analysis.R
\name{pkern_kmeans}
\alias{pkern_kmeans}
\title{Cluster 1-dimensional coordinates onto grid lines of unknown position and number}
\usage{
pkern_kmeans(x, nmax = NA, dmin = NA, nstart = 25, cw = 1)
}
\arguments{
\item{x}{numeric vector, the positions of points to cluster}

\item{nmax}{positive integer, the maximum desired number of x grid lines}

\item{dmin}{numeric, lower limit on distance between grid lines}

\item{nstart}{positive integer, passed to stats:kmeans}

\item{cw}{non-negative numeric, the complexity exponent (see details)}
}
\value{
names list of: grid line positions "grid", and "input" their index in x
}
\description{
This function uses heuristics to find a best fitting irregular grid for the values
in \code{x}, with up to \code{nmax} grid lines and minimum separation distance \code{dmin}. It
does a brute force search of all \code{1:nmax} candidates for the number of grid lines,
using kmeans clustering (with \code{nstart} randomly assigned starting values) to find
the k grid line positions. Tuning parameter \code{cw} can be increased to favor simpler
grids.
}
\details{
Different values of k are ranked according to two cost functions

E(k) = sum of squared distances between points and their grid lines
C(k) = 1 + sum of squared differences between the expected and actual occupancy of grid line

where if grid line i has ni points assigned, its actual occupancy is (n_i) / (n/k), and
the "expected" occupancy is that of a complete grid, or n/k. C(k) will be higher for
incomplete grids, particularly ones where high k causes grid lines to be assigned to
solitary points. The score function for k is the sum:

S(k) = k + log( E(k) ) + \code{cw} * log( C(k) )

This is evaluated for all k = 1,...\code{nmax} and the grid line set with smallest S(k) is
returned. If \code{nmax} is not supplied, it is set to the smaller of: n, or \code{4*sqrt(n)},
or the number of unique coordinate positions. If \code{dmin} is not supplied, it is set
to \code{(range(x)/n)/4}
}
\examples{
nx = 35
ny = 24
coords = expand.grid(seq(nx), seq(ny)) + rnorm(nx*ny, 0, 1/5)
plot(coords)
n = nrow(coords)
x = coords[,1] \%>\% as.vector
pkern_kmeans(x)
}
