% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkern_pars.R
\name{pkern_kp}
\alias{pkern_kp}
\title{Return named vector of kernel parameters initialized to NA}
\usage{
pkern_kp(k)
}
\arguments{
\item{k}{character, the kernel name, one of 'exp', 'gau', 'sph', 'gxp', 'mat'}
}
\value{
named vector of NAs, a placeholder for kernel parameters
}
\description{
Convenience function for looking up the number of parameters for a given
kernel, and their names. Returns a vector of NAs that can be used as a
placeholder in kernel definition lists.
}
\examples{
pkern_kp('mat')
}
