% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkern_covariance.R
\name{pkern_kplot}
\alias{pkern_kplot}
\title{Make a heatmap of covariances around a grid's central point}
\usage{
pkern_kplot(pars, dims = c(101, 101))
}
\arguments{
\item{pars}{list of two parameter lists recognized by \code{pkern_corr} and named entry "v"}

\item{dims}{c(nx, ny), the number of x and y lags to include}
}
\value{
returns the RasterLayer plotted by the function
}
\description{
This function visualizes a separable kernel by building a grid of size \code{dims}
and coloring cells according to their covariance with the central cell.
}
\details{
if either entry of \code{dims} is an even number, it incremented by 1 in order to
make the definition of "central" unambigious.
}
\examples{
pars = list(x=pkern_corr('mat'), y=pkern_corr('mat'))
pkern_kplot(pars)
pkern_kplot(pars, dims=c(10,10))
}
