% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkern_computation.R
\name{pkern_kprod}
\alias{pkern_kprod}
\title{Left-multiplication with a Kronecker product: evaluates (X (x) Y) * vector}
\usage{
pkern_kprod(X, Y, z, trans = FALSE)
}
\arguments{
\item{X}{matrix or Matrix object, the x-component matrix}

\item{Y}{matrix or Matrix object, the y-component matrix}

\item{z}{numeric vector, the vector to left-multiply}

\item{trans}{logical, indicating the inputs \code{X} and \code{Y} are transposed}
}
\value{
TODO
}
\description{
Evaluates a Kronecker product times a vector by the standard trick of representing
the vector as a matrix M and evaluating \code{Y \%*\% M \%*\% t(X)}.
}
\details{
For computational efficiency, the function computes the result by a composition
of cross-products with the transposed input component matrices. The argument
\code{trans=TRUE} indicates that \code{t(X)} and \code{t(Y)} have been passed to \code{X} and \code{Y},
so that the transpose step (which is slow for large matrices) can be skipped.
}
\examples{
# TODO
}
