% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkern_index.R
\name{pkern_mat2vec}
\alias{pkern_mat2vec}
\title{Column-vectorization indices}
\usage{
pkern_mat2vec(ij, ni, simplified = TRUE)
}
\arguments{
\item{ij}{n x 2 matrix, the row and column indices}

\item{ni}{number of rows in the matrix}

\item{simplified, }{if FALSE, the function returns an n x 1 matrix}
}
\value{
a vector of indices in the vectorized system
}
\description{
Maps matrix indices to vectorization indices. The function returns the vector
indices associated with the supplied matrix (i, j) indices.
}
\details{
Column vectorization (as in base::as.vector) builds a length(mn) vector by stacking
the columns of an m X n matrix, with the leftmost column appearing first in the vector
and the rightmost column last. Matrix element i,j gets mapped to element \code{(i + m*(j-1))}
in the vector. This function returns that index.

\code{ij} can be a matrix or a list of length-n vectors "i" and "j" (in that order), or a vector
representing a single point at the given row and column number. \code{ni} can be a vector of the
form \code{c(ni, nj)} (the return value of \code{dim} for example) in which case its first element
(the number of rows) is used.
}
\examples{

# get column-vectorized ordering of points from matrix indices
ni = 10
gyx = expand.grid(i=seq(ni), j=seq(ni))
pkern_mat2vec(gyx, ni)
}
