% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkern_model.R
\name{pkern_nLL}
\alias{pkern_nLL}
\title{Compute negative log-likelihood for parameter vector \code{p}}
\usage{
pkern_nLL(
  p,
  g_obs,
  pars_fix,
  X = 0,
  iso = FALSE,
  quiet = TRUE,
  log_scale = FALSE
)
}
\arguments{
\item{p}{numeric vector of covariance parameters accepted by \code{pkern_pars_update}}

\item{g_obs}{list of form returned by \code{pkern_grid} (with entries 'gdim', 'gres', 'gval')}

\item{pars_fix}{list of form returned by \code{pkern_pars} (with entries 'y', 'x', 'eps', 'psill')}

\item{X}{numeric, vector, matrix, or NA, the mean or its linear predictors, passed to \code{pkern_LL}}

\item{iso}{logical, indicates to use identical kernels for x and y (\code{pars$x} is ignored)}

\item{quiet}{logical indicating to suppress console output}
}
\value{
numeric, the negative log-likelihood of \code{p} given \code{g_obs}
}
\description{
Returns the negative log-likelihood of parameter vector \code{p} for the covariance
model \code{pars_fix}, given data grid \code{g_obs}.
}
\details{
This is a wrapper for \code{-pkern_LL()} allowing parameters to be passed as a numeric
vector instead of a list (for use in optimization etc). parameters in \code{p} are copied
to \code{pars_fix} and passed to the likelihood computer.

\code{p} is the vector of covariance parameters to test. It must be in the form (length,
order) accepted by \code{pkern_pars_update(pars_fix, p, na_omit=TRUE, iso=iso)}.
}
\examples{
# set up example grid and data
g_obs = pkern_grid(10)
g_obs$gval = rnorm(10^2)

# get some default parameters and vectorize them
pars_fix = pkern_pars(g_obs, 'gau')
p = pars_fix |> pkern_pars_update()
pkern_nLL(p, g_obs, pars_fix)
pkern_nLL(p, g_obs, pars_fix, eps_scaled=TRUE)

}
