% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkern_estim.R
\name{pkern_optim}
\alias{pkern_optim}
\title{Fit covariance parameters to data by maximum (profile) likelihood using optim}
\usage{
pkern_optim(
  g_obs,
  pars_fix = "gau",
  X = 0,
  iso = F,
  control = list(),
  quiet = F,
  log_scale = TRUE,
  method = "L-BFGS-B",
  ...
)
}
\arguments{
\item{g_obs}{list of form returned by \code{pkern_grid} (with entries 'gdim', 'gres', 'gval')}

\item{pars_fix}{list of fixed kernel parameters, with NAs indicating parameters to fit}

\item{X}{numeric, vector, matrix, or NA, the mean or its linear predictors, passed to \code{pkern_LL}}

\item{iso}{logical, indicating to constrain the y and x kernel parameters to be the same}

\item{control}{list, passed to \code{stats::optim}}

\item{quiet}{logical, indicating to suppress console output}

\item{...}{named arguments to pass to \code{pkern_bds}}
}
\value{

}
\description{
Fit covariance parameters to data by maximum (profile) likelihood using optim
}
\examples{
# set up example grid and data
g_obs = pkern_grid(10)
g_obs$gval = rnorm(10^2)
pkern_optim(g_obs, quiet=TRUE)

# repeat with one or more parameters fixed
pars_fix = pkern_pars_make('gau') # NA parameters list
pars_fix$psill = 1
pkern_optim(g_obs, pars_fix, quiet=TRUE)
pars_fix$y$kp = 1
pkern_optim(g_obs, pars_fix, quiet=TRUE)

# iso mode constrains x parameters to equal y parameters
pkern_optim(g_obs, iso=T, quiet=TRUE)
pkern_optim(g_obs, pars_fix, iso=T, quiet=TRUE)

}
