% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkern_computation.R
\name{pkern_precompute}
\alias{pkern_precompute}
\title{Compute important indexing vectors relating to a subgrid within a larger grid}
\usage{
pkern_precompute(gdim, gli, pars, zobs = NULL, makev = FALSE)
}
\arguments{
\item{gdim}{integer vector c(ny, nx) the size of the outer grid (in number of grid lines)}

\item{gli}{list of integer vectors "y" and "x", indexing the grid lines that form the subgrid}

\item{pars}{list of "y" and "x" kernel parameters}

\item{zobs}{(optional) vector, containing NAs where data are missing in the subgrid}

\item{makev}{logical, indicates to compute variance matrix components for points off subgrid}
}
\value{
a large list of vectors and matrices
}
\description{
Regular subgrids have computationally convenient properties that we can exploit
when predicting onto a supergrid
}
\details{
TODO: flesh out documentation
(cross) covariance matrices

The length of \code{obs} should equal the product of the lengths of the grid line vectors in
\code{gxy}, and its elements should be in column-vectorized order. Only the locations of the NAs
(and not the other contents) of \code{obs} matter.
}
\examples{
# TODO
}
