% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkern_computation.R
\name{pkern_qf}
\alias{pkern_qf}
\title{efficient quadratic form computer: \code{t(x) \%*\% M \%*\% y}}
\usage{
pkern_qf(M, x, y)
}
\arguments{
\item{M}{numeric matrix}

\item{x}{numeric vector}

\item{y}{numeric vector}
}
\value{
TODO
}
\description{
efficient quadratic form computer: \code{t(x) \%*\% M \%*\% y}
}
\examples{
# TODO
}
