% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkern_indexing.R
\name{pkern_r2c}
\alias{pkern_r2c}
\title{Swap grid indices from row-vectorized to column-vectorized order}
\usage{
pkern_r2c(
  dims,
  in.byrow = TRUE,
  out.byrow = FALSE,
  flipx = FALSE,
  flipy = FALSE
)
}
\arguments{
\item{dims}{c(nx, ny), the number of x and y grid lines in source grid}

\item{in.byrow}{logical indicating if source indexing is row-vectorized}

\item{out.byrow}{logical indicating if destination indexing is row-vectorized}

\item{flipx}{logical indicating to flip the grid left-to-right}

\item{flipy}{logical indicating to flip the grid top-to-bottom}
}
\value{
a reordering for the vectorized grid. ie a permutation of seq(prod(dims))
}
\description{
\code{pkern} uses column-vectorization with y decreasing, to mirror standard mathematical
notation for matrices. This utility function is for converting to and from this ordering.
}
\details{
dims can be a length-2 vector (as described below) or a RasterLayer object, from which
the dimensions are extracted. In the latter case
}
\examples{
dims = c(12, 13)
permutation = pkern_r2c(dims, in.byrow=TRUE, out.byrow=FALSE)
matrix(permutation, nrow=dims[1], byrow=TRUE) |> as.vector()
}
