% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkern_indexing.R
\name{pkern_r45}
\alias{pkern_r45}
\title{Rotate a rectangular array by 45 degrees clockwise}
\usage{
pkern_r45(z, dims = NULL)
}
\arguments{
\item{z}{either a numeric matrix or its length-\code{prod(dims)} column-vectorized version}

\item{dims}{c(nx, ny) the number of x and y grid lines}
}
\value{
either a list ("z", "dims") or matrix
}
\description{
performs the rotation f(x,y) = (x + y, -x + y) about the center of a rectangular
array (45 degrees counterclockwise), with distances scaled by sqrt(2)/2 to snap
grid cells to those of a larger array.
}
\details{
When \code{z} is a matrix, \code{dims} is ignored (with a warning) and can be omitted, and
the function returns the new rotated array . When \code{z} is a vector, it should be
in column-vectorized order, and the function returns a list containing the dimensions
of the larger array (nx, ny), and the column-vectorized data.

NAs for all output grid points not mapped to \code{z}
}
