% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkern_raster.R
\name{pkern_rasterize}
\alias{pkern_rasterize}
\title{Snap points to a subgrid in a RasterLayer}
\usage{
pkern_rasterize(pts, g, regular)
}
\arguments{
\item{pts, }{numeric vector or n x 2 matrix (with columns for x and y) of point coordinates}

\item{g, }{RasterLayer}

\item{regular, }{logical, indicating to select a regular set of grid lines}
}
\value{
a RasterLayer containing the data from \code{pts}
}
\description{
Snap points to a subgrid in a RasterLayer
}
