% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkern_indexing.R
\name{pkern_regular}
\alias{pkern_regular}
\title{Snap an irregular subgrid to nearest regular version}
\usage{
pkern_regular(gid, ng = max(gid), sep = NA)
}
\arguments{
\item{gid}{grid line numbers of the subgrid (sorted in ascending order)}

\item{ng}{positive integer no less than \code{max(gid)}, the number of grid lines in full grid}

\item{sep}{integer between 1 and \code{(ng-1)}, desired resolution, in number of grid lines}
}
\value{
vector of same length as \code{gid} containing the new (snapped) grid line numbers
}
\description{
Grid line numbers \code{gid} should be a subset of \code{seq(ng)}. The function finds the
nearest regular subset, in terms of the sum of squared differences. ie it maps
the input \code{gid} to a new subset of \code{seq(ng)} in which subsequent elements are all
separated by an equal number of grid lines (\code{sep}). If \code{sep} is not supplied, it
is estimated as the median of the differences among the \code{gid}.
}
\details{
Note that input grid lines will be merged when doing so results in a smaller sum of
squares.
}
\examples{
ng = 100
gid = ( seq(1, ng, 5) + rnorm(20, 0, 1) ) |> ceiling() |> sort()
diff(gid)
gid.snap = pkern_regular(gid)
diff(gid.snap)
}
