% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkern_index.R
\name{pkern_rescale}
\alias{pkern_rescale}
\title{Up-scale or down-scale a grid}
\usage{
pkern_rescale(g, up = NULL, down = NULL, vals = TRUE)
}
\arguments{
\item{g}{any object accepted or returned by \code{pkern_grid}}

\item{up}{integer > 0, an up-scaling factor}

\item{down}{integer > 0, a down-scaling factor}

\item{vals}{logical, indicates to return data vector as well as grid info}
}
\value{
a grid list of the form returned by \code{pkern_grid}
}
\description{
Changes the resolution of a grid by a factor of (positive integer) \code{up} or \code{down}.
}
\details{
This constructs a new grid of based on \code{g}, where: if \code{up} is supplied, a sparser
(up-scaled) version of \code{g} is returned, with every \code{up}th grid line sampled; and
if \code{down} is supplied, a denser version of \code{g} is returned where every \code{down}th
grid line is copied over from \code{g} (and the rest initialized to NA).

This effects the scaling of resolution (\code{g$gres}) by \code{up} (or \code{1/down}). It does
not do any interpolation to fill new grid-points created with \code{up} and it does not
change the bounding box of \code{g}.
}
\examples{

# example data
gdim = c(50, 53)
g = pkern_grid(gdim)
gval = pkern_sim(g, modifyList(pkern_pars(g), list(eps=0)), quiet=TRUE)
g_obs = modifyList(g, list(gval=gval))
pkern_plot(g_obs)

# upscale
pkern_plot(pkern_rescale(g=g_obs, up=1)) # does nothing
pkern_plot(pkern_rescale(g=g_obs, up=2))
pkern_plot(pkern_rescale(g=g_obs, up=3))

# downscale
pkern_plot(pkern_rescale(g=g_obs, down=1)) # does nothing
pkern_plot(pkern_rescale(g=g_obs, down=2))
pkern_plot(pkern_rescale(g=g_obs, down=3))

# turn vals off to get grid configuration only
pkern_plot(g)
pkern_plot(pkern_rescale(g=g_obs, up=2, vals=FALSE))
pkern_plot(pkern_rescale(g=g_obs, down=2, vals=FALSE))

}
