% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkern_covariance.R
\name{pkern_rho}
\alias{pkern_rho}
\title{Solve for range parameter given a correlation value}
\usage{
pkern_rho(cval, pars, d, upper = 1000 * d)
}
\arguments{
\item{cval}{numeric between 0 and 1, the correlation}

\item{pars}{a kernel parameter list recognized by \code{pkern_corr}}

\item{d}{positive numeric, the distance at which to evaluate the kernel}

\item{upper}{positive numeric, an upper limit for the search}
}
\value{

}
\description{
Solve for range parameter given a correlation value
}
