% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkern_model.R
\name{pkern_sim}
\alias{pkern_sim}
\title{Generate random draw from multivariate normal distribution for grid points}
\usage{
pkern_sim(
  g_obs,
  pars = pkern_pars(g_obs),
  fac = NULL,
  add_frac = 1e-16,
  quiet = FALSE
)
}
\arguments{
\item{pars}{list, covariance parameters in form returned by \code{pkern_pars}}

\item{fac}{(optional) list, precomputed factorization of component correlation matrices}

\item{add_frac}{numeric > 0, small constant for adjusting negative eigen-values (see details)}

\item{quiet}{logical, indicating to suppress warnings about negative eigen-values}

\item{g}{any object accepted or returned by \code{pkern_grid}}
}
\value{
numeric vector, the vectorized grid data
}
\description{
Generate random draw from multivariate normal distribution for grid points
}
\examples{

# example grid and covariance parameters
gdim = c(25,15)
g = pkern_grid(gdim)
pars = pkern_pars(g, 'mat')

# this example has a large nugget effect
gval = pkern_sim(g, pars)
pkern_plot(matrix(gval, gdim))

# plot with yx coordinates
g_sim = modifyList(g, list(gval=gval))
pkern_plot(g_sim)

# repeat without nugget effect for smooth field
pars0 = modifyList(pars, list(eps=0))
gval = pkern_sim(g, pars0)
pkern_plot(matrix(gval, gdim))

# plot with yx coordinates
g_sim = modifyList(g, list(gval=gval))
pkern_plot(g_sim)

# parameters are automatically assigned if not supplied
pkern_pars(g) # default is gau x gau with ranges roughly equal to grid side lengths
pkern_plot(matrix(pkern_sim(g), gdim))

}
