% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkern_analysis.R
\name{pkern_snap_grid}
\alias{pkern_snap_grid}
\title{Snap a set of points to a larger regular grid by least Manhattan distance}
\usage{
pkern_snap_grid(coords, gxy)
}
\arguments{
\item{coords}{n X 2 matrix of "x" and "y" coordinates or sf object}

\item{gxy}{list of two numeric vectors, the grid line positions "x" and "y"}
}
\value{

}
\description{
Snap a set of points to a larger regular grid by least Manhattan distance
}
