% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkern_snap.R
\name{pkern_snap_plot}
\alias{pkern_snap_plot}
\title{Plots a set of 1 or 2d points and their snapped versions}
\usage{
pkern_snap_plot(g, pts = NULL, ppars = list())
}
\arguments{
\item{g}{numeric vector, matrix, dataframe or list, the grid line coordinates (see details)}

\item{pts}{numeric vector, matrix, dataframe or list, the point coordinates (see details)}

\item{ppars}{list, (optional) plotting parameters "gcol", "mcol", "dcol", "pcol", and "yaxis"}
}
\value{
draws a plot and returns nothing
}
\description{
The function plots the points in \code{pts} over the grid lines in \code{g}, with the option of
also drawing connecting lines between points and grid intersections, according to a
given mapping.
}
\details{
This is mainly intended for visualizing the output of \code{pkern_snap} and related functions.

\code{pts} can be a vector of 1d positions (in which case the plot shows input order against
position), or a matrix whose columns are the "y" and "x" coordinates (in that order,
unless they are named), or a data frame (which is coerced to matrix), or a list of
coordinate vectors.

\code{g} can be a numeric vector, or list of two (in "y", "x" order), or a list containing the
grid line locations (as vector or list) in element "g", and, optionally,  "map" which
provides a mapping for drawing connecting lines. If supplied, "map" should be an integer
vector (or list of two) with length(s) matching (those of) \code{pts}, and with elements
indexing grid line vectors, ie from the sequence \code{1:length(g$g)}

When \code{g$g} but not \code{g$map} is supplied, it is assumed that \code{g$g} and \code{pts} have
the same length and ordering, and lines are drawn connecting their respective elements.
When \code{g} is a list without named element "g", it is assumed to be a list of grid line
vectors (and no connecting lines are drawn).

Different classes of \code{pts} and \code{g} (and "map") can be mixed but they must have the
same dimensionality; ie a 2d points set \code{pts} must be passed with a 2d grid line object
\code{g}, and if "map" is supplied, it must be a list of two mapping vectors.

\code{ppars} is a list with any of the following (optional) named elements: "gcol", "mcol",
"dcol", and "pcol". They are the colors of, respectively, the grid lines, the snapped grid
lines connecting lines, and points (passed as parameter "col" to the appropriate \code{graphics}
call); and "yaxis" is a logical that applies only in the 1d case, indicating to plot point
locations along the y (instead of x) axis.
}
\examples{
# vector g (1d case)
g = seq(1, 100, by=2)
ng = length(g)
pts = g + stats::rnorm(ng)
pkern_snap_plot(g, pts)

# plot again with with default snapping (assumes g and pts have same order)
pkern_snap_plot(g=list(g=g), pts)

# plot again with a random mapping
pkern_snap_plot(list(g=g, map=sample(seq(ng), ng)), pts)

# matrix g (2d case)
gdim = c(30, 25)
sep = c(2, 3)
g = Map(\(d, s) seq(1, d, by=s), d=gdim, s=sep)
xy = expand.grid(g)
pts = xy + stats::rnorm(prod(dim(xy)))
pkern_snap_plot(g, pts)

# plot again with a mapping from pts to g
pkern_snap_plot(g=list(g=g), pts)

# TODO: examples with pkern_snap
}
