% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkern_snap.R
\name{pkern_snap_plot_defaults}
\alias{pkern_snap_plot_defaults}
\title{Set defaults for pmap and pts in pkern_snap_plot}
\usage{
pkern_snap_plot_defaults(gyx, pmap, pts)
}
\arguments{
\item{gyx}{numeric vector of grid-line coordinates, or two such vectors (y, x) in a list}

\item{pmap}{integer vector of mapping indices, or two such vectors (y, x) in a list, or NULL}

\item{pts}{integer vector of point coordinates, or two such vectors (y, x) in a list, or NULL}
}
\value{
list containing \code{pmap}, \code{pts}, and a logical indicating if input was 1d
}
\description{
Helper function for pkern_snap_plot
}
\details{
Parses the input arguments \code{gyx}, \code{pmap} and \code{pts}, returning a modified
version of \code{pmap} and \code{pts} with defaults set as needed, along with the
dimension.
}
