% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkern_snap.R
\name{pkern_snap_score_1d}
\alias{pkern_snap_score_1d}
\title{Compute squared snapping distances (and map) from 1d points to a 1d regular sub-grid}
\usage{
pkern_snap_score_1d(gl, pts, sep = 1, off = NULL, mapping = FALSE)
}
\arguments{
\item{gl}{numeric vector, the grid line coordinate values in 1 dimension}

\item{pts}{numeric vector, the point coordinates in 1 dimension}

\item{sep}{integer vector of length 2, the sub-grid spacing indices}

\item{off}{positive integer, the sub-grid offset index, an element of \code{seq(sep)}}

\item{mapping}{logical, indicating to return extra information in a list}
}
\value{
numeric, the total squared snapping distance, or list (see DETAILS)
}
\description{
This function constructs the sub-grid of \code{g} having spacing \code{sep} (ie \code{sep-1} grid
points lie between adjacent sub-grid points) and computes the sum of the squared
distances between elements of \code{pts} and their nearest (snapped) sub-grid point.
}
\details{
When \code{mapping=TRUE} the function returns the individual squared snapping distances
for each point, along with information about the mapping from grid to points, in
a named list containing:

gyx: numeric, the input grid-line coordinates (1d)
d2total: numeric, the total squared snapping distance
pmap: numeric vector, the element of \code{g} mapped to each point in \code{pts}
ngl: integer, the minimum dimension of the sub-grid containing all snapped points
noc: integer, the number of occupied sub-grid lines (can be less than ngl)
}
\examples{
sep = 3
ng = 1e2
g = seq(ng)
sg = seq(sample(sep, 1), ng, by=sep)

# sample part of the subgrid and add noise
npts = round( (3/4) * length(sg) )
pts = sort(sample(sg, npts)) + stats::rnorm(npts)
pkern_snap_plot(g, pts)

# snap to subgrid and plot result
snap = pkern_snap_score_1d(gl=sg, pts, sep=1, mapping=TRUE)
pkern_snap_plot(snap, pts)

# snap to outer grid and plot result
snap = pkern_snap_1d(g, pts, 1)
pkern_snap_plot(snap, pts)

}
