% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkern_raster.R
\name{pkern_toRaster}
\alias{pkern_toRaster}
\title{Convert column-vectorized grid to SpatRaster}
\usage{
pkern_toRaster(g = NULL, gdim = NULL, template = NULL)
}
\arguments{
\item{g}{either the column-vectorized data or a list containing it (in element "gval")}

\item{gdim}{c(ny, nx), the number rows and columns in the full grid}

\item{template}{optional RasterLayer to use as template (for setting crs, resolution etc)

Converts a column-vectorized vector or matrix to a SpatRaster, or, if terra is unavailable,
a RasterLayer.

\code{g} can either be a matrix (in which case \code{gdim} can be omitted) or its vectorized
equivalent, or a list containing elements returned by \code{pkern_fromRaster}. If the list
has an element named "gdim", it replaces any argument supplied to \code{gdim}. Other elements
of the list (eg 'crs') are passed on to the RasterLayer construction call as needed.

When template is supplied, it supercedes any properties (eg 'crs') supplied in \code{rvec}.}
}
\value{
a RasterLayer containing the data from \code{rvec}
}
\description{
Convert column-vectorized grid to SpatRaster
}
\examples{
if( requireNamespace('raster') ) {
r.in = system.file('external/rlogo.grd', package='raster') |> raster::raster(band=1)
pkern.in = pkern_fromRaster(r.in)
r = pkern_toRaster(pkern.in)
print(r)
}
}
