% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkern_plot.R
\name{pkern_toString}
\alias{pkern_toString}
\title{Extract kernel parameters as plot-friendly strings}
\usage{
pkern_toString(pars, nsig = 3)
}
\arguments{
\item{pars}{a kernel parameters list ("k", "kp"), or list of two of them ("x", "y")}
}
\value{
a character string or list (depending on the contents of \code{pars}, see details)
}
\description{
Generates a list of useful strings based on the kernel names and parameter values
in \code{pars} for plot labels and such. \code{pars} should either specify kernel(s) by name
or else be a parameters list of the form accepted by \code{pkern_corr}.
}
\details{
When \code{pars} is a character string, the function parenthesizes it with square brackets
and returns the resulting string. When \code{pars} is a vector of character strings, it
collapses the two parenthesized strings into a single vector separated by an "x".

When \code{pars} is a list of kernel parameters for a single dimension, the function
returns a named list of two character strings: the kernel name ("k") formatted as
above, and the parameter(s) ("kp") (also square bracketed, in "name = value" format
where "value" is rounded to \code{nsig} significant digits).

When \code{pars} is a list of two such kernel parameter lists (named "x" and "y"), the
function again returns a list with elements "k" and "kp", containing the strings from
each component collapsed with a " x " delimiter (indicating direct product). The third
element, "main", is a language object containing the bolded kernel name, and (if "nug"
is supplied) a fourth element supplies a string "nug" with the nugget effect in
parentheses.
}
\examples{
kname = 'mat'
pkern_toString(kname)
pkern_toString(rep(kname, 2))
pkern_toString(list(k=kname))

pars = pkern_corr(kname)
pkern_toString(pars)

kname = c('mat', 'sph')
pars = pkern_corr(kname)
pkern_toString(pars)
pkern_toString(modifyList(pars, list(nug=0.5)))

}
