% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkern_variograms.R
\name{pkern_tvario}
\alias{pkern_tvario}
\title{Theoretical semivariances along x or y direction}
\usage{
pkern_tvario(pars, v, d = NULL)
}
\arguments{
\item{pars}{list of kernel parameters, in form recognized by \code{pkern_corr}}

\item{v}{numeric, the pointwise variance}

\item{d}{vector of nonegative numeric spatial lags to evaluate}
}
\value{
either a vector the same length as \code{d}, or an anonymous function of distance
}
\description{
Convenience function for getting theoretical semivariance values for a set of
spatial lags, assuming (intrinsic) stationarity.
}
\details{
Kernel parameters \code{pars} are passed to the correlation function (\code{pkern_corr})
to get correlation values at distances \code{d}. The result is converted to semivariance
assuming pointwise variance \code{v} and returned as a vector the same length as \code{d}.

When \code{d} is NULL, the function instead returns the semivariance function as an
anonymous function of distance.
}
\examples{
pars.mat = pkern_corr('mat')
pkern_tvario(pars.mat, v=1, 1:10)
sv = pkern_tvario(pars.mat, v=1)
sv(1:10)
}
