% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkern_variograms.R
\name{pkern_tvg}
\alias{pkern_tvg}
\title{Theoretical variogram function}
\usage{
pkern_tvg(pars, d = NULL)
}
\arguments{
\item{pars}{list of kernel parameters (see details)}

\item{d}{vector of nonegative numeric spatial lags to evaluate}
}
\value{
either a vector the same length as \code{d}, or an anonymous function of distance
}
\description{
Returns either the variogram function for the (intrinsically stationary) process
specified in kernel parameter list \code{pars}; or, if distances \code{d} are supplied, the
vector of variogram values evaluated at \code{d}.
}
\details{
\code{pars} and \code{d} are passed to the correlation function (\code{pkern_corr}) to get the
correlogram values \code{C(d)/C(0)} (where \code{C} is the covariance function), and the result
is converted to the variogram by the relation,

\code{2*gamma(d) = nug + 2 * C(0) * (1  - C(d)/C(0))},

where the nugget variance \code{nug} (by default 0) is specified in \code{pars$nug} and the
partial sill (\code{C(0)}, by default 1) in \code{pars$psill}.

When \code{pars} is a 2d separable kernel (ie when it has elements "y" and "x" recognized
by \code{pkern_corr}), \code{d} should be a list or matrix containing the "y" and "x" distances
(vectors \code{x} and \code{y}). The correlogram \code{C(d)/C(0)} is then formulated as
\verb{Cx(x) * Cy(y) ) / C(0)}, where \code{Cx} and \code{Cy} are the 1-dimensional correlograms for
the kernels \code{pars$x} and \code{pars$y}.

When \code{d} is \code{NULL}, the function returns an anonymous function of distance (the
variogram). Otherwise \code{d} should be a numeric vector of distances, in which case the
function returns the variogram evaluated at those distances.
}
\examples{
d = seq(1e2)/10
pars = pkern_corr('mat')
pkern_tvg(pars)
pkern_tvg(pars, d=d)

vg = pkern_tvg(pars)
plot(d, vg(d))
}
