% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkern_covariance.R
\name{pkern_unpack}
\alias{pkern_unpack}
\title{Vectorize x and y component parameters}
\usage{
pkern_unpack(x, y, p = NULL)
}
\arguments{
\item{x}{list of x component}

\item{sig}{}
}
\value{

}
\description{
Vectorize x and y component parameters
}
