% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkern_indexing.R
\name{pkern_unwhich}
\alias{pkern_unwhich}
\title{Inverse for base::which}
\usage{
pkern_unwhich(idx, n, asinteger = FALSE)
}
\arguments{
\item{idx}{vector of positive integers no greater than n}

\item{n}{positive integer}
}
\value{
length-n vector
}
\description{
base::which takes a logical vector and returns the indices of TRUE elements.
This function does the opposite, creating a length-n logical vector that is
TRUE at the indices in idx, and FALSE otherwise
}
\examples{
n = 25
foo = sample(c(TRUE, FALSE), size=n, replace=TRUE)
idx.foo = which(foo)
pkern_unwhich(idx.foo, n)
}
