% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkern_computation.R
\name{pkern_variance}
\alias{pkern_variance}
\title{Pointwise variance of kriging predictions}
\usage{
pkern_variance(pc, quiet = FALSE, idx = NULL)
}
\arguments{
\item{pc}{a list of precomputed objects returned by \code{pkern_cmean}}

\item{quiet}{logical indicating to suppress progress bar on console}

\item{idx}{(optional) integer vector indexing a subset of eigenvectors}
}
\value{
TODO
}
\description{
Computes the conditional variance of kriging predictions on a grid
}
\details{
TODO: write up the proof for this method and write up a short description
here.
}
\examples{
# TODO
}
