% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkern_variograms.R
\name{pkern_vario}
\alias{pkern_vario}
\title{Directional sample variograms for regular 2D gridded data}
\usage{
pkern_vario(
  gdim,
  vec,
  dmax = NA,
  fit.method = "rmedian",
  diagonal = TRUE,
  gres = NA,
  quiet = FALSE,
  nmax = 10000,
  vcalc = TRUE
)
}
\arguments{
\item{gdim}{vector c(ny, nx) of positive integers, the number of grid lines}

\item{vec}{numeric vector of data, in column vectorized order (length \code{prod(dims)})}

\item{dmax}{numeric, the maximum lag distance (only shorter lags are sampled)}

\item{fit.method}{character, one of "matheron", "rmean", "rmedian", passed to \code{pkern_xvario}}

\item{diagonal}{logical, if TRUE, semivariogram results are returned also for diagonals}

\item{gres}{length-2 vector of positive numeric, the spacing distance of y and x grid lines}

\item{quiet}{logical, indicating to suppress console output}

\item{nmax}{positive integer, the maximum number of point pairs to process in each sample}

\item{vcalc}{logical, indicates to compute the sample variance (which can be relatively slow)}
}
\value{
list of \code{pkern_xvario} output "x", "y" (and "d1", "d2"); sample variance "v"; and "gres"
}
\description{
A wrapper for calls to \code{pkern_xvario} to get the directional sample variograms
from regular lattice data along the y and x dimensions, and optionally along the
diagonals.
}
\details{
The function makes the appropriate calls to \code{pkern_xvario} and returns the results
in list elements "y", "x". If \code{diagonal=TRUE}, the function also returns "d1" and
"d2", the y and x semivariograms after a 45 degree counterclockwise rotation of the
input grid (see \code{pkern_r45}) assuming equal y and x resolution. For unequal resolution
the rotation angle is \code{atan(gres[2]/gres[1])}.

For convenience, \code{gdim} can be a list containing both the 'gdim' and 'gres' arguments.
If the sub-list \code{gdim$sg} exists, the function copies these arguments from \code{gdim$sg$gdim}
and \code{gdim$sg$gres}. Otherwise it looks for them in \code{gdim$gdim} and \code{gdim$gres}.
For example, \code{gdim} can be the output of \code{pkern_snap} (to sample from snapped points) or
\code{pkern_fromRaster} (to sample from raster values).

You can (and, with large grids, probably should) set \code{dmax} to specify a maximum
sampling distance. This both speeds computations, and omits pathological tail behavior
from variogram fitting downstream. \code{dmax} should be given in the same units as \code{gres}.
Output lags are also given in these units, as are the x-axis values drawn by
\code{pkern_vario_plot}.

When \code{gres} is not supplied (directly, or in \code{gdim}), the function assumes
that the x and y grid lines have spacing 1.
}
\examples{
gdim = c(1e2, 2e2)
pars.1d = pkern_corr('gau')
pars.2d = list(y=pars.1d, x=pars.1d, psill=1)
vec = pkern_sim(pars.2d, gdim)

# sample first 10 lags and plot
vario = pkern_vario(gdim, vec, lags=seq(10), quiet=TRUE)
pkern_vario_plot(vario, pars.2d)

# example with unequal resolution and multiple replicates
gres = c(2,1)
nrep = 10
vec = lapply(seq(nrep), \(i) pkern_sim(pars.2d, gdim, gres=gres, makeplot=FALSE))
vario = pkern_vario(gdim, vec, lags=seq(10), gres=gres, quiet=TRUE)
pkern_vario_plot(vario, pars.2d)

}
