% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkern_variograms.R
\name{pkern_vario_fit}
\alias{pkern_vario_fit}
\title{Fit a theoretical separable covariance model to a separated empirical semivariogram}
\usage{
pkern_vario_fit(
  vario,
  xpars = "exp",
  ypars = xpars,
  v = NA,
  nug = NA,
  nmin = 2,
  ninitial = 1
)
}
\arguments{
\item{vario}{list, the return value from a call to \code{pkern_vario}}

\item{xpars}{character or list, recognizable (as \code{pars}) by \code{pkern_corr}}

\item{ypars}{character or list, recognizable (as \code{pars}) by \code{pkern_corr}}

\item{v}{length-3 numeric vector, c('min', 'ini', 'max'), bounds and initial value for variance}

\item{nug}{length-3 numeric vector, c('min', 'ini', 'max'), bounds and initial value for nugget}

\item{nmin}{positive integer, the minimum number of samples to include a lag}

\item{ninitial}{positive integer, the number of initial value sets to test (see details)}
}
\value{

}
\description{
Pipe the results of a \code{pkern_vario} call to this function to estimate model
parameters for \code{xpars} and \code{ypars}, the x and y component kernels, by weighted
least squares (Cressie, 2015), where \code{stats::optim} ("L-BFGS-B" method with default
settings is used to solve the minimization problem.
}
\details{
\code{xpars} and \code{ypars} may be specified as character strings (the exponential
"exp" is the default), in which case they are assigned suggested initial values
and bounds. If parameter lists (recognized by \code{pkern_corr}), are supplied, the
function uses its "kp" entry for initial values, and if the list contains entries
"lower" and/or "upper", they are used as bounds.

The variance is estimated simultaneously with the correlation kernel parameters
using bounds and initial value in \code{v}. If these are not supplied, the function
sets reasonable defaults.
}
