% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkern_sample.R
\name{pkern_vario_fun}
\alias{pkern_vario_fun}
\title{Theoretical variogram function}
\usage{
pkern_vario_fun(pars, d = NULL)
}
\arguments{
\item{pars}{list of the form returned by \code{pkern_pars} with entries 'y', 'x', 'eps', 'psill'}

\item{d}{numeric vector or list with vector entries 'y' and 'x', the distances to evaluate}
}
\value{
vector (for list \code{d}) or data.frame (for vector \code{d})
}
\description{
Computes the value of the semi-variogram function defined by covariance model \code{pars}
at the component (y and x) lags in list \code{d}.
}
\details{
If \code{d} is a list, its 'y' and 'x' components should be equal length nonnegative numeric
vectors. The function returns the semi-variance in a vector of the same length.

If \code{d} is a numeric vector, it is interpreted as a set of distances to at which to
evaluate the range of the (anisotropic) semi-variogram function. The function returns
a list with named numeric vectors 'min' and 'max'.
}
\examples{
gdim = c(10, 15)
d_max = sqrt(sum(gdim^2))
pars = pkern_pars(gdim, 'mat')

d_test = seq(0, d_max, length.out=1e2)
vario = pkern_vario_fun(pars, d=list(y=0, x=d_test))
plot(d_test, vario)
vario_lims = pkern_vario_fun(pars, d=d_test)
lines(d_test, vario_lims[,1])
lines(d_test, vario_lims[,2])

}
