% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkern_variograms.R
\name{pkern_vario_plot}
\alias{pkern_vario_plot}
\title{Plot 1-dimensional empirical semivariograms for a separable model}
\usage{
pkern_vario_plot(vario, pars = NULL, nmin = 1, ptitle = NULL, shade = TRUE)
}
\arguments{
\item{vario}{list of semivariance data, the return value of \code{pkern_vario} or \code{pkern_xvario}}

\item{pars}{list of kernel parameters including variance in named element "v"}

\item{nmin}{positive integer, the minimum number of point pairs needed to include in plot}

\item{ptitle}{character or vector of two, title(s) for the plot(s)}

\item{shade}{logical, indicating to shade points according to the number of samples}
}
\value{
prints a plot and returns nothing
}
\description{
Plot 1-dimensional empirical semivariograms for a separable model
}
