% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkern_indexing.R
\name{pkern_vec2mat}
\alias{pkern_vec2mat}
\title{Inverse column-vectorization indices}
\usage{
pkern_vec2mat(idx, ni, out = "matrix")
}
\arguments{
\item{idx}{a vector of positive integers}

\item{ni}{number of rows in the matrix}

\item{out}{either 'matrix' or 'list'}
}
\value{
a two column matrix of integers (row and column numbers) with \code{length(idx)} rows
}
\description{
This function is for doing the inverse of \code{as.vector(amatrix)}. It returns the row and
column numbers (i,j) associated with a given vector element after column-vectorization.
}
\details{
If a grid size vector \code{c(ni, nj)} is passed to \code{ni}, the function uses its first element.
}
\examples{

# show how elements are ordered in `base::matrix`
ni = 5
nj = 6
matrix.indexing = matrix(1:prod(ni, nj), ni)
print(matrix.indexing)
as.vector(matrix.indexing)

# doing the inverse
pkern_vec2mat(2, ni)
pkern_vec2mat(c(1,2,7), ni)
}
