% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkern_covariance.R
\name{pkern_bds}
\alias{pkern_bds}
\title{Suggest bounds for parameters of y and x component kernels for a separable grid model}
\usage{
pkern_bds(pars, gres = NA, nmax = 1, cmin = 0.05, cmax = 0.5)
}
\arguments{
\item{pars}{kernel parameter list (recognized by \code{pkern_corr}) or list of two of them}

\item{gres}{vector c(dy, dx) or positive numeric, the distances between adjacent grid lines}

\item{nmax}{positive integer, the number of grid cells between "distant" points}

\item{cmin}{positive numeric, minimum allowable correlation between adjacent points}

\item{cmax}{positive numeric, maximum allowable correlation between "distant" points}
}
\value{
kernel parameter list containing "kp", "lower", "upper", or list of two such lists
}
\description{
\code{pars} should be a kernel parameter list, or a list of two of them (named "y", "x").
The function adds vectors "kp" (initial values), "lower" and "upper" (bounds), wherever
they are missing. As a shortcut, \code{pars} can also be a character string (or vector of two)
naming the desired kernel(s), in which case the function sets a suggested initial value
(see \code{pkern_corr}).
}
\details{
Bounds for shape parameters are hard-coded, whereas the bounds for the range parameter
are determined based on the grid layout (\code{gres}, \code{gdim}) and the arguments \code{cmin}, \code{cmax}.
These are, respectively, the minimum correlation between adjacent grid points, and the
maximum correlation between points at a distance of \code{nmax} cells apart. The corresponding
range parameter values are then estimated using \code{pkern_rho}, which does a grid search over
the permissible shape parameters (if there are any).

Initial values for the range and shape parameters, when missing from \code{pars}, are set to
the midpoints of their permissible ranges.

By default \code{cmin} is set to 0.05, so that the effective range of the kernel is bounded
below by the shortest interpoint distance. The default for \code{cmax} was selected by trial
and error.
}
\examples{

# suggested bounds depend on grid resolution and nmax
pars = 'mat'
pkern_bds(pars)
pkern_bds(pars, gres=100)
pkern_bds(pars, gres=100, nmax=100)

# specify separable 2D kernels with a vector of two names
pars = c('exp', 'mat')
pkern_bds(pars)

# incomplete parameter definition lists are filled in
pars = list(k='mat', upper=c(10,10))
pkern_bds(pars)

}
