% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkern_pars.R
\name{pkern_bds}
\alias{pkern_bds}
\title{Set default parameter covariance parameter bounds}
\usage{
pkern_bds(
  pars,
  g,
  var_obs = NULL,
  w = 5,
  rows = NULL,
  initial = NULL,
  lower = NULL,
  upper = NULL
)
}
\arguments{
\item{pars}{list or character vector of 1-2 kernel names (see \code{pkern_pars})}

\item{g}{list, a pkern grid definition (or object accepted by \code{pkern_grid})}

\item{var_obs}{positive numeric, the sample variance of data \code{g$gval}}
}
\value{
a data frame of intial values and lower/upper bounds for the parameters in \code{pars}
}
\description{
Returns a data-frame of initial values and upper/lower bounds on covariance
parameters for the kernel names in \code{pars}.
}
\details{
psill (partial sill) and eps (measurement variance) are assigned minima (and
maxima) equal to 1/100th times (and 100 times) the sample variance \code{var_obs}.
The user must supply either \code{var_obs} or \code{g$gval} so it can be computed.

Each range parameter (rho for y and x) is assigned a minimum (and maximum) equal to
the shortest (and longest) inter-point distance along the corresponding dimension;
its initial value is set to the geometric mean of the upper and lower bounds.
Distances are computed using \code{g$gres} and its (element-wise) product with \code{g$gdim}.

Shape parameter bounds are hard-coded, and are set conservatively to avoid problems
with numerical precision in functions like \code{exp} and \code{gamma} when evaluating very
large or small distances.
}
\examples{
gdim = c(10, 15)
z = prod(gdim) |> rnorm()
g = pkern_grid(gdim) |> modifyList(list(gval=z))
pkern_bds('mat', g)
pkern_bds('mat', g, lower=0)
pkern_bds('mat', g, rows=c('eps', 'psill'), lower=c(0, 0.5))
}
