% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkern_covariance.R
\name{pkern_bds}
\alias{pkern_bds}
\title{Suggest bounds for parameters of y and x component kernels for a separable grid model}
\usage{
pkern_bds(pars, ds = NA, cmin = 0.05, cini = 0.9, cmax = cini + (1 - cini)/2)
}
\arguments{
\item{pars}{kernel parameter list (recognized by \code{pkern_corr}) or list of two of them}

\item{ds}{vector c(dy, dx) or positive numeric, the distances between adjacent grid lines}

\item{cmin}{positive numeric, minimum allowable correlation between adjacent grid points}

\item{cini}{positive numeric, initial correlation between adjacent grid points}

\item{cmax}{positive numeric, maximum allowable correlation between adjacent grid points}
}
\value{
kernel parameter list containing "kp", "lower", "upper", or list of two such lists
}
\description{
\code{pars} should be a kernel parameter list, or a list of two of them (named "y", "x").
The function adds vectors "kp" (initial values), "lower" and "upper" (bounds), wherever
they are missing. As a shortcut, \code{pars} can also be a character string (or vector of two)
naming the desired kernel(s), in which case the function sets a suggested initial value
(see \code{pkern_corr}).
}
\details{
Bounds for shape parameters are hard-coded, whereas the bounds for the range parameter
are determined based on the grid resolution (\code{ds}, the distance between grid lines in
y and x directions) and a user supplied correlation range; \code{cmin} is the minimum allowable
correlation between adjacent grid points, and \code{cmax} is the maximum. The corresponding "rho"
values are then estimated using \code{pkern_rho} (via a grid search over the permissible
shape parameters, if there are any)

Initial values for the shape parameters are set to the midpoints of their permissible
ranges. Initial value for the range parameters are set such that adjacent cells have
correlation \code{cini}

By default \code{cmin} is set to 0.05, so that the effective range of the kernel is bounded
below by the shortest interpoint distance. The default for \code{cmax} is the midpoint between
\code{cini} and 1. \code{cini} is by default set to
}
\examples{

# suggested bounds depend on grid resolution
pars = 'mat'
pkern_bds(pars)
pkern_bds(pars, ds=100)

# specify separable 2D kernels with a vector of two names
pars = c('exp', 'mat')
pkern_bds(pars)

# initial values are set based on lower and upper
pars = list(k='mat', upper=c(10,10))
pkern_bds(pars)

}
