% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkern_computation.R
\name{pkern_cmean}
\alias{pkern_cmean}
\title{Compute conditional mean of points on a grid given a sample taken over a subgrid}
\usage{
pkern_cmean(zobs, gdim, pars, gli = NULL, pc = FALSE)
}
\arguments{
\item{zobs}{numeric vector of observed data, possibly with NAs}

\item{gdim}{integer vector (ny, nx), the size of the full grid}

\item{pars}{list of kernel parameter lists "y" and "x", nugget "nug", and distance scaling "ds"}

\item{gli}{list of two integer vectors, indexing the y and x grid lines of the regular subgrid}

\item{pc}{either logical (default FALSE), or a list of precomputed objects (see details)}
}
\value{
TODO
}
\description{
Given a separable spatial covariance model, \code{pars}, and sample data \code{zobs} on a
a regular subgrid \code{gli} of the grid of dimensions \code{gdim}, the function computes the
expected value of all n (\code{prod(gdim)}) points in the full grid.
}
\details{
Tthe full n-dimensional covariance matrix (V) is represented using Kronecker products
(eg. as described in Gilboa et al., 2015; and Koch et al., 2020), dramatically speeding
up calculations and reducing memory demands.

\code{zobs} should be supplied in column-vectorized order, and its length must equal the
product of the lengths of the vectors in \code{gli} (ie the number of points in the subgrid).
Unsampled points in the subgrid should be indicated by NAs in \code{zobs}.

The vectors in \code{gli} indicate the grid line numbers that make up the subgrid - ie they
should be subsets of \code{seq(gdim[1])} and \code{seq(gdim[2])}, in ascending order, and must be
regularly spaced.

Nugget variance (in \code{pars$nug}) is added to the diagonal of the eigenvalue matrix in
the eigendecomposition of V. Note that a small nugget effect will often solve computational
issues arising from numerically singular covariance matrices.
}
\examples{
# TODO
}
