% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkern_index.R
\name{pkern_coords}
\alias{pkern_coords}
\title{Return coordinates of a grid of points in column-vectorized order}
\usage{
pkern_coords(g, out = "matrix", corner = FALSE)
}
\arguments{
\item{g}{any object accepted by \code{pkern_grid}}

\item{out}{character indicating return value type, either 'list', 'matrix', or 'sf'}

\item{corners}{logical, indicates to only return the corner points}
}
\value{
a matrix, list, or sf POINT collection, in column vectorized order
}
\description{
Expands a set of y and x grid line numbers in the column-vectorized order returned
by \code{pkern_grid}.
}
\details{
This is similar to \code{base::expand.grid} but with the first dimension (y) descending
instead of ascending.

\code{out='sf'} returns an \code{sf} simple features object containing points in the same order,
with data (if any) copied from \code{g$gval} into column 'gval'. Note that \code{prod(g$gdim)}
points are created, which can be slow for large grids.
}
\examples{
gdim = c(5,3)
g_example = pkern_grid(list(gdim=gdim, gres=c(0.5, 0.7), gval=seq(prod(gdim))))
pkern_coords(g_example)
pkern_coords(g_example, out='list')

# corner points
pkern_coords(g_example, corner=TRUE)
pkern_coords(g_example, corner=TRUE, out='list')

# sf output type
if( requireNamespace('sf') ) {

# make the points
sf_coords = pkern_coords(g_example, out='sf')

# data are copied to variable 'gval'
plot(sf_coords, pch=16)

}

}
