% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkern_covariance.R
\name{pkern_corr}
\alias{pkern_corr}
\title{Stationary 1D correlation kernels}
\usage{
pkern_corr(pars, d = NA)
}
\arguments{
\item{pars}{list with elements "k", the kernel name, and "p" the parameter vector}

\item{d}{numeric vector of length n, the distances to evaluate}
}
\value{
length-n vector, or a matrix of parameter bounds, or a parameter list (see details)
}
\description{
Computes stationary correlation function values for (nonegative) 1-dimensional
distances \code{d}. Parameter list entry \code{pars$kp} supplies the kernel parameter(s),
where \code{pars$k} is one of the following kernel names
}
\details{
"exp": exponential (special case of "gex" with p=1 and \code{d<-abs(d)}). \code{kp = rho}
"gau": gaussian/stable (special case of "gex" with p=2). \code{kp = rho}
"sph": spherical (AKA stable/Gaussian for p=2). \code{kp = rho}
"gex": gamma-exponential. \code{kp = c(rho, p)}
"mat": Whittle-Matern ((Handcock and Wallis parameterization). \code{kp = c(rho, kap)}

If any element of \code{d} is NA, the function ignores \code{d} and instead returns a list
containing the parameter names, along with suggested initial values and bounds to
use in optimization.

If instead of a list, \code{pars} is one of the kernel name strings, the function returns
an example of the parameter list format for the kernel, with suggested initial values
(and \code{d} is ignored)
}
\examples{
pars.mat = pkern_corr('mat')
pars.mat
pkern_corr(pars.mat)
pkern_corr(pars.mat, d=1:10)
}
