% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkern_var.R
\name{pkern_corr}
\alias{pkern_corr}
\title{Stationary 1D correlation kernels}
\usage{
pkern_corr(pars, d = NA)
}
\arguments{
\item{pars}{list with elements "k", the kernel name, and "kp" the parameter vector}

\item{d}{numeric vector of length n, the distances to evaluate}
}
\value{
length-n vector or a list of parameters and bounds (see details)
}
\description{
Computes stationary correlation function values for the n (nonegative) 1-dimensional
distances in \code{d}. Parameter list entry \code{pars$kp} supplies the kernel parameter(s).
}
\details{
\code{pars$k} must be one of the following kernel names:

"exp": exponential (special case of "gex" with shape "p"=1)
"gau": gaussian/stable (special case of "gex" with shape "p"=2)
"sph": spherical (AKA stable/Gaussian for p=2)

"gex": gamma-exponential (with shape "p")
"mat": Whittle-Matern (Handcock and Wallis parameterization, with shape "kap")

where the first three have 1 single range parameter, and the last two have both a
range and shape parameter.

For the 1-parameter kernels, \code{pars$kp} is the range parameter value ("rho" ); For the
2-parameter kernels, \code{pars$kp} is a vector whose first element is "rho", and second
element is the shape parameter ("p" or "kap"). Note that names in \code{pars$kp} are ignored
and only the order matters - the range parameter always comes first.
}
\examples{

# define a 1D kernel and grab values
pars = pkern_pars(10)[['x']]
pkern_corr(pars, d=1:10)

}
