% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkern_covariance.R
\name{pkern_corr}
\alias{pkern_corr}
\title{Stationary 1D correlation kernels}
\usage{
pkern_corr(pars, d = NA)
}
\arguments{
\item{pars}{list with elements "k", the kernel name, and "kp" the parameter vector}

\item{d}{numeric vector of length n, the distances to evaluate}
}
\value{
length-n vector or a list of parameters and bounds (see details)
}
\description{
Computes stationary correlation function values for the n (nonegative) 1-dimensional
distances in \code{d}. Parameter list entry \code{pars$kp} supplies the kernel parameter(s).
}
\details{
\code{pars$k} must be one of the following kernel names:

(1 parameter)

"exp": exponential (special case of "gex" with shape "p"=1)
"gau": gaussian/stable (special case of "gex" with shape "p"=2)
"sph": spherical (AKA stable/Gaussian for p=2)

(2 parameters)

"gex": gamma-exponential (with shape "p")
"mat": Whittle-Matern (Handcock and Wallis parameterization, with shape "kap")

For the 1-parameter kernels, \code{pars$kp} is the range parameter value ("rho" ); For the
2-parameter kernels, \code{pars$kp} is a vector whose first element is "rho", and second
element is the shape parameter ("p" or "kap"). Note that names in \code{pars$kp} are ignored
and only the order matters - the range parameter must come first.

When \code{d} is NA (or contains at least one NA value), or if instead of a list \code{pars} is
a kernel name string (or a vector of two of them), then \code{d} is ignored and the function
passes \code{pars} to \code{pkern_bds} and returns the result. This sets suggested parameter values
and bounds wherever they are missing from \code{pars}.
}
\examples{

# define a 1D kernel and grab values
pars = list(k='mat', kp=c(10, 2))
pkern_corr(pars, d=1:10)

# get suggested bounds
pkern_corr(pars)
pkern_corr('mat')

}
