% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkern_snap.R
\name{pkern_estimate_sep}
\alias{pkern_estimate_sep}
\title{Estimate separation distance between grid lines}
\usage{
pkern_estimate_sep(g, pts, distinct = TRUE, bias = 0, dlim = NULL)
}
\arguments{
\item{g}{vector of grid line locations in ascending order}

\item{pts}{vector of point locations in 1D}

\item{distinct}{logical, indicating to look for 1-to-1 mappings}

\item{bias}{nonegative numeric, a tuning parameter}

\item{dlim}{length-2 positive numeric, lower/upper bounds for separation distance}
}
\value{
the estimated value of \code{sep}
}
\description{
Estimates the subgrid resolution best matching an irregular set of points.
Helper function for \code{pkern_snap_1d}
TODO: flesh out documentation here and tidy inner loop in code
}
\examples{
sep = 25
ng = 10*sep
g = seq(ng)
sg = seq(sample(sep, 1), ng, by=sep)

# sample part of the subgrid and add noise
npts = round( (3/4) * length(sg) )
pts = sort(sample(sg, npts)) + stats::rnorm(npts)
pkern_snap_plot(g, pts)

# estimate separation and plot result
sep = pkern_estimate_sep(g, pts)
sep
snap = pkern_snap_1d(g, pts, sep)
pkern_snap_plot(snap, pts)

# add duplication
ndupe = 10
pts = sapply(pts, \(p) p + stats::rnorm(ndupe)) |> c()

# fit again
pkern_snap_plot(g, pts)
sep = pkern_estimate_sep(g, pts, distinct=FALSE)
sep
snap = pkern_snap_1d(g, pts, sep=sep, distinct=FALSE)
pkern_snap_plot(snap, pts)

}
