% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkern_computation.R
\name{pkern_fit}
\alias{pkern_fit}
\title{Fit a spatial covariance model by numerical maximum likelihood}
\usage{
pkern_fit(
  zobs,
  gsnap,
  ypars = "gau",
  xpars = ypars,
  v = NULL,
  nug = NULL,
  add = 0,
  control = list()
)
}
\arguments{
\item{zobs}{numeric vector or matrix of subgrid data (with NAs at unsampled points)}

\item{gsnap}{list of vectors, output from \code{pkern_snap}}

\item{ypars}{character or list, recognizable (as \code{pars}) by \code{pkern_corr}}

\item{xpars}{character or list, recognizable (as \code{pars}) by \code{pkern_corr}}

\item{v}{length-3 numeric vector, c('min', 'ini', 'max'), bounds and initial value for variance}

\item{nug}{length-3 numeric vector, c('min', 'ini', 'max'), bounds and initial value for nugget}

\item{add}{positive integer, add additional initial parameter sets are tested (see details)}
}
\value{

}
\description{
Fit a spatial covariance model by numerical maximum likelihood
}
\examples{
#
}
