% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkern_raster.R
\name{pkern_fromRaster}
\alias{pkern_fromRaster}
\title{Vectorize a RasterLayer in column-vectorized order}
\usage{
pkern_fromRaster(r, what = "all")
}
\arguments{
\item{r}{a RasterLayer to vectorize}

\item{what}{character, one of: 'gdim', 'ds', 'u', 'crs', 'yx', 'values' or 'all'}
}
\value{
For default \code{what=='all'}, a named list containing 6 objects (see details)
}
\description{
Calls to raster::values or raster::getValues produce the data in row-vectorized
order. This function reorders them to column-vectorized order so they can be
passed to other pkern_* functions
}
\details{
The return value depends on argument \code{what}:

"gdim": returns integer vector \code{c(i, j)}, the number of y and x grid lines
"ds": returns \code{c(dy, dx)}, the x and y distances between grid lines (ie the resolution)
"crs": returns the CRS code for the raster (if any)
"yx": returns \code{list(y, x)}, a list of vectors containing the y and x grid line positions
"values": returns a vector containing the raster data in column-vectorized order
"all": returns a named list containing the above three objects
}
\examples{
if( requireNamespace('raster') ) {
r.in = system.file('external/rlogo.grd', package='raster') |> raster::raster(band=1)
pkern.in = pkern_fromRaster(r.in)
pkern_plot(pkern.in)
pkern.in$values |> matrix(pkern.in$gdim) |> pkern_plot()
}
}
