% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkern_raster.R
\name{pkern_fromRaster}
\alias{pkern_fromRaster}
\title{Vectorize a RasterLayer in column-vectorized order}
\usage{
pkern_fromRaster(r, omode = "list")
}
\arguments{
\item{r}{a RasterLayer to vectorize}

\item{omode}{character indicating return value type, either 'list' or 'vector'}
}
\value{
either the data vector or a list containing it along with the dimensions
}
\description{
Calls to raster::values or raster::getValues produce the data in row-vectorized
order. This function reorders them to column-vectorized order, optionally returning
the dimensions in (ny, nx) format (the reverse of what you get with dim(raster))
}
\examples{
require(raster)
r.in = system.file('external/rlogo.grd', package='raster') |> raster(band=1)
pkern.in = pkern_fromRaster(r.in)
pkern.in$data |> matrix(ncol=pkern.in$dims[1]) |> image()
pkern_toRaster(pkern.in, template=r.in)
pkern_toRaster(pkern_fromRaster(r.in, omode='vector'), template=r.in)
}
