% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkern_indexing.R
\name{pkern_idx_sg}
\alias{pkern_idx_sg}
\title{Find column-vectorized index of a subgrid}
\usage{
pkern_idx_sg(dims, gx = NULL, gy = NULL)
}
\arguments{
\item{dims}{c(nx, ny), the number of x and y grid lines in the full grid}

\item{gx}{vector of positive integers no greater than nx, the x grid lines of the subgrid}

\item{gy}{vector of positive integers no greater than ny, the y grid lines of the subgrid}
}
\value{
integer vector indexing the subgrid points with respect to the full grid
}
\description{
A grid with nx columns and ny rows has grid lines gx=1:nx and gy=1:ny.
A subgrid includes only a subset of these grid lines. This function
uses the grid line numbers to identifies points on the subgrid by their
column-vectorized index with respect to the full grid
}
\details{
This is Useful when taking a subset of a dataframe or submatrix of a
covariance matrix for gridded data in column-vectorized order.

gx or gy NA indicates to use all grid lines. For convenience, list(gx, gy)
can supplied in argument gx, but only when gy is NA
}
\examples{
dims = c(5,6)
pkern_idx_sg(dims, c(1,3,5), c(2, 4))
}
