% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkern_plot.R
\name{pkern_kplot}
\alias{pkern_kplot}
\title{Make a heatmap of covariances/correlations around a grid's central point}
\usage{
pkern_kplot(pars, dims = NULL, smoothed = FALSE, maxx = 200, ppars = list())
}
\arguments{
\item{pars}{list of two parameter lists "x" and "y", and (optionally) "v", "nug"}

\item{dims}{c(nx, ny), the number of x and y lags to include}

\item{smoothed}{logical, whether to smooth the image (passed to \code{pkern_plot})}

\item{maxx}{integer or integer vector, plotting resolution (passed to \code{pkern_plot})}

\item{ppars}{list of plotting arguments (passed to \code{pkern_plot})}
}
\value{
returns nothing but prints a contour plot of the specified kernel
}
\description{
This function displays a separable kernel by building a grid of size \code{dims},
assigning to each cell the covariance (or correlation) with the central cell,
then passing the result to \code{graphics::image} for visualization.
}
\details{
\code{pars} should be a list containing elements "x" and "y", which are lists of parameters
for the x and y component kernels (recognized by \code{pkern_corr}). Optionally, \code{pars} can
also include a (positive numeric) pointwise variance term in "v". Covariance parameters
from both kernels are printed as a subtitle, with values rounded to 3 decimal places.

When \code{dims} is omitted, it is set to the smallest size such that the plot shows the
effective range of the kernel components in both the x and y directions. if either
entry of \code{dims} is an even number, it incremented by 1 in order so that "central"
is  unambiguous.

If \code{v} and \code{nug} are not supplied, plot labels are changed to show "correlation".
}
\examples{
pars = pkern_bds(c('gxp', 'mat'))
pkern_kplot(pars)
pkern_kplot(pars, dims=c(50,30))
pkern_kplot(c(pars, list(v=2, nug=0.5)), dims=c(50,30))
pkern_kplot(c(pars, list(v=2, nug=0.5)), dims=c(50,30), smoothed=TRUE)
}
