% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkern_plot.R
\name{pkern_kplot}
\alias{pkern_kplot}
\title{Make a heatmap of covariances/correlations around a grid's central point}
\usage{
pkern_kplot(pars, gdim = NULL, gres = 1, ppars = list())
}
\arguments{
\item{pars}{list of two parameter lists "y" and "x", and (optionally) "v", "nug"}

\item{gdim}{c(ni, nj), the number of y and x lags to include}

\item{gres}{c(dy, dx), the y and x distances between adjacent grid lines}

\item{ppars}{list of plotting arguments (passed to \code{pkern_plot})}
}
\value{
returns nothing but prints a contour plot of the specified kernel
}
\description{
This function displays a separable kernel by building a grid of size \code{gdim},
assigning to each cell the covariance (or correlation) with the central cell,
then passing the result to \code{graphics::image} for visualization. Covariance
parameters from both kernels are printed as a subtitle, with values rounded
to 3 decimal places.
}
\details{
\code{pars} should be a list containing elements "y" and "x", which are lists of parameters
for the y and x component kernels (recognized by \code{pkern_corr}). Optionally, \code{pars} can
also include a (positive numeric) pointwise variance and nugget effect (numerics "v"
and "nug"). If \code{pars} contains an element named "gres", it supercedes any argument to \code{gres}.
If \code{v} and \code{nug} are supplied, the legend title is set to "covariance", otherwise it is
set to "correlation".

When \code{gdim} is omitted, it is set to the smallest size such that the plot shows the
effective range of the kernel components in both the y and x directions. if either
entry of \code{gdim} is an even number, it incremented by 1 so that "central point"
is unambiguous.
}
\examples{
pars = pkern_bds(c('gxp', 'mat'))
pkern_kplot(pars)
gdim = c(30,50)
pkern_kplot(pars, gdim)
pkern_kplot(c(pars, list(psill=2, nug=0.5)), gdim)
pkern_kplot(c(pars, list(psill=2, nug=0.5)), gdim, ppars=list(smoothed=TRUE))
}
