% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkern_computation.R
\name{pkern_LL}
\alias{pkern_LL}
\title{Likelihood function for sampled points in a subgrid}
\usage{
pkern_LL(pars, zobs, sgdim, pc = FALSE)
}
\arguments{
\item{pars}{kernel parameter list, see \code{pkern_corr}}

\item{zobs}{numeric vector or matrix of subgrid data (with NAs at unsampled points)}

\item{sgdim}{integer vector, the size of the subgrid (ny, nx), required if \code{zobs} is a vector}

\item{pc}{logical or list of matrices, precomputed objects (see details)}
}
\value{

}
\description{
\code{zobs} can either be a matrix (in which case \code{sgdim}, if supplied, must match \code{dim(zobs)})
or a vector of length \code{prod(sgdim)}.
}
\examples{
# TODO
}
