% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkern_indexing.R
\name{pkern_mat2vec}
\alias{pkern_mat2vec}
\title{Column-vectorization indices}
\usage{
pkern_mat2vec(ij, ni, simplified = TRUE)
}
\arguments{
\item{ij}{nx2 matrix, the row (y) and column (x) indices}

\item{ni}{number of rows in the matrix}

\item{simplified, }{if FALSE, the function returns an n x 1 matrix}
}
\value{
a vector of indices in the vectorized system
}
\description{
Returns the index in the vector associated with the supplied i, j values
}
\details{
Column vectorization (as in base::as.vector) builds a length(mn) vector by stacking
the columns of an m X n matrix. The (i,j)th element of the matrix is mapped to the
\code{(i + m*(j-1))}th element of the vector.

\code{ij} can be a matrix or a list of length-n vectors "i" and "j" (in that order), or a vector
representing a single point at the given row and column number. \code{ni} can be a vector
c(ni, nj), in which case its first element is selected.
}
\examples{

# get column-vectorized ordering of points from matrix indices
ni = 10
g = expand.grid(i=seq(1, ni), j=seq(1:ni))
pkern_mat2vec(g, ni)
}
