% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkern_indexing.R
\name{pkern_mat2vec}
\alias{pkern_mat2vec}
\title{Column-vectorization indices}
\usage{
pkern_mat2vec(ij, ny)
}
\arguments{
\item{ij}{n x 2 matrix of column (x) and row (y) indices}

\item{ny}{number of rows in the matrix}
}
\value{
a vector of indices in the vectorized system
}
\description{
Column vectorization (as in base::as.vector) builds a length(mn) vector by stacking
the columns of an m X n matrix. The (i,j)th element of the matrix is mapped to the
\code{(i + m*(j-1))}th element of the vector. This function returns the index in the vector
associated with the supplied i, j values
}
\details{
If \code{ij} has named columns 'x' and 'y', these are interpreted as 'j' and 'i'. If column
names 'i' and 'j' also appear, these supercede the columns 'x' and 'y'. If no column names
are supplied, the function assumes they are in order 'j', 'i'
}
