% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkern_computation.R
\name{pkern_precompute}
\alias{pkern_precompute}
\title{Compute important indexing vectors relating to a subgrid within a larger grid}
\usage{
pkern_precompute(
  gdim,
  map,
  pars = NULL,
  zobs = NULL,
  makev = FALSE,
  sparse = TRUE
)
}
\arguments{
\item{gdim}{integer vector c(ni, nj) the size of the outer grid}

\item{map}{list of integer vectors "i" and "j", row and column indices of the subgrid}

\item{pars}{(optional) list of "x" and "y" kernel parameters}

\item{zobs}{(optional) vector, containing NAs where data are missing in the subgrid}

\item{makev}{logical, indicates to compute variance matrix components for points off subgrid}

\item{sparse}{logical, indicates to use Matrix::Matrix representations of large matrices}
}
\value{
a large list of vectors and matrices
}
\description{
Regular subgrids have computationally convenient properties that we can exploit
when predicting onto a supergrid
}
\details{
TODO: flesh out documentation
(cross) covariance matrices

The length of \code{obs} should equal the product of the lengths of the grid line vectors in
\code{gxy}, and its elements should be in column-vectorized order. Only the locations of the NAs
(and not the other contents) of \code{obs} matter.
}
\examples{
# TODO
}
