% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkern_indexing.R
\name{pkern_r45}
\alias{pkern_r45}
\title{Rotate a rectangular array by 45 degrees clockwise}
\usage{
pkern_r45(z, gdim = NULL)
}
\arguments{
\item{z}{either a numeric matrix, its column-vectorization, a RasterLayer, or a list}

\item{gdim}{integer vector c(ni, nj), the number of rows and columns in the grid}
}
\value{
a numeric matrix containing the rotated array
}
\description{
performs the rotation f(x,y) = (x + y, -x + y) about the center of a rectangular
array, rotating the grid 45 degrees clockwise with distances scaled by 2 so that
coordinates snap to those of a larger array.
}
\details{
NAs are assigned to all output grid points not mapped to \code{z}.

When \code{z} is an object containing the grid size (matrix, RasterLayer or list),
\code{gdim} is ignored (with a warning) and can be omitted. When \code{z} is a vector, it
should be in column-vectorized order and have length \code{prod(gdim)}.
}
\examples{

# a wide example modified from `graphics::.filled.contour`
gdim = c(73,50)
y = seq(-pi, pi, length.out=gdim[1])
x = seq(-pi, pi, length.out=gdim[2])
d = sqrt(outer(y^2, x^2, '+'))
z = c( sin(d^2) * exp(-d) )
pkern_plot(z, gdim)
pkern_plot( pkern_r45(z, gdim) )

# a tall example with raster
if( requireNamespace('raster') ) {
r = system.file('external/rlogo.grd', package='raster') |> raster::raster(band=1)
rg = pkern_fromRaster(r)
pkern_plot(rg)
pkern_plot( pkern_r45(rg) )
}
}
