% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkern_covariance.R
\name{pkern_rho}
\alias{pkern_rho}
\title{Solve for range parameter given a correlation value}
\usage{
pkern_rho(cval, pars, d = 1, upper = 1000 * d)
}
\arguments{
\item{cval}{numeric between 0 and 1, the correlation}

\item{pars}{a 1D kernel parameter list (with elements "k", "kp") recognized by \code{pkern_corr}}

\item{d}{positive numeric, the distance at which to evaluate the kernel}

\item{upper}{positive numeric, an upper limit for the search}
}
\value{
the estimated "rho" value producing semivariance \code{cval} at distance \code{d}
}
\description{
Solve for range parameter given a correlation value
}
\examples{
pars = pkern_corr('exp')
cval = 0.5
pkern_rho(cval, pars, d=1)
pkern_rho(cval, pars, d=10)
pkern_corr(utils::modifyList(pars, list(kp=pkern_rho(0.5, pars, d=10))), d=10)

}
