% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkern_grid.R
\name{pkern_snap}
\alias{pkern_snap}
\title{Title}
\usage{
pkern_snap(from, g = NULL, vals = TRUE, crop = TRUE)
}
\arguments{
\item{from}{matrix, data frame, or points object from \code{sp} or \code{sf}, the source points}

\item{g}{any grid object accepted or returned by pkern_grid, the destination grid}

\item{vals}{logical, indicating to copy data values from points to the grid}
}
\value{

}
\description{
Title
}
\examples{
# 3x2 example and sf output type

pkern_coords(c(3,2))

if( requireNamespace('sf') ) {


g = pkern_grid(list(gdim=c(13, 12), gres=0.7))
g$gyx$x = g$gyx$x + 10
g$gyx$y = g$gyx$y - 100

eg_sf = pkern_coords(c(13, 12), out='sf')

from = eg_sf


from = as(eg_sf,'Spatial')






}

}
