% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkern_indexing.R
\name{pkern_snap}
\alias{pkern_snap}
\title{Snap a set of (possibly irregular) points to a larger grid}
\usage{
pkern_snap(pts, g, regular = FALSE)
}
\arguments{
\item{pts, }{input coordinates to snap}

\item{g, }{grid lines of larger grid}

\item{regular}{}
}
\value{
list
}
\description{
Snap a set of (possibly irregular) points to a larger grid
}
\examples{
resx = 10
resy = 10
nx = 25
ny = 30
xy = list(resx*seq(nx), resy*seq(ny))
coords = expand.grid(xy) + rnorm(nx*ny)
gxy.snap = pkern_snap(pts=coords, g=xy, regular=F)
plot(coords)
abline(v = gxy.snap$x$gval)
abline(h = gxy.snap$y$gval)
}
