% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkern_snap.R
\name{pkern_snap_plot}
\alias{pkern_snap_plot}
\title{Plots a set of 1 or 2d points and their snapped versions}
\usage{
pkern_snap_plot(gyx, pts = NULL, ppars = list())
}
\arguments{
\item{gyx}{numeric vector, matrix, dataframe or list, the grid line coordinates (see details)}

\item{pts}{numeric vector, matrix, dataframe or list, the point coordinates (see details)}

\item{ppars}{list, (optional) plotting parameters "gcol", "mcol", "dcol", "pcol", and "yaxis"}
}
\value{
draws a plot and returns nothing
}
\description{
The function plots the points in \code{pts} over the grid lines in \code{gyx}, with the option of
also drawing connecting lines between points and grid intersections to show snapping
distances.
}
\details{
This is mainly intended for visualizing the output of \code{pkern_snap} and related functions.

\code{pts} can be a vector of 1d positions (in which case the plot shows input order against
position), or a matrix whose columns are the "y" and "x" coordinates (in that order,
unless they are named), or a data frame (which is coerced to matrix), or a list of
coordinate vectors, or an \code{sf} points object.

\code{gyx} can be a numeric vector in the 1d case; or list of two (in "y", "x" order); or a
list containing the grid line locations (as vector or list) in element "gyx", and,
optionally, "pmap" which provides a mapping for drawing connecting lines. For example
you can pass the return value of \code{pkern_snap} as \code{gxy}

If supplied, \code{gyx$pmap} should be an integer vector (or list of two) with length(s)
matching those in \code{pts} - its elements index grid lines, so they should have values
drawn (possibly with repitition) from the sequences \code{seq_along(gyx$gyx$y)} and
\code{seq_along(gyx$gyx$x)}

When \code{gyx$gyx} but not \code{gyx$pmap} is supplied, it is assumed that \code{gyx$gyx} and \code{pts} have
the same length and ordering, and lines are drawn connecting their respective elements.
When \code{gyx} is a list without named element "gyx", it is assumed to be a list of grid line
vectors (and no connecting lines are drawn).

\code{ppars} is a list with any of the following (optional) named elements: "gcol", "mcol",
"dcol", and "pcol". They are the colors of, respectively, the grid lines, the snapped grid
lines connecting lines, and points (passed as parameter "col" to the appropriate \code{graphics}
call).
}
\examples{
# 1d case
gyx = seq(1, 100, by=2)
pkern_snap_plot(gyx)
ng = length(gyx)
pts = gyx + stats::rnorm(ng)
pkern_snap_plot(gyx, pts)

# plot again with with points mapped to grid line that generated them
pkern_snap_plot(gyx=list(gyx=g, pmap=seq(gyx)), pts)

# plot again with a random mapping
pkern_snap_plot(list(gyx=gyx, pmap=sample(seq(ng), ng)), pts)

# 2d case
gdim = c(30, 25)
sep = c(2, 3)
gyx = Map(\(d, s) seq(1, d, by=s), d=gdim, s=sep)
yx = expand.grid(gyx)
pts = yx + stats::rnorm(prod(dim(yx)))
pkern_snap_plot(gyx, pts)

# plot again with with points mapped to grid line that generated them
pmap = Map(\(p, s) match(p, s), p=yx, s=gyx)
pkern_snap_plot(gyx=list(gyx=gyx, pmap=pmap), pts)

# TODO: examples with pkern_snap
}
