% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkern_raster.R
\name{pkern_toRaster}
\alias{pkern_toRaster}
\title{Convert column-vectorized grid to RasterLayer}
\usage{
pkern_toRaster(rvec = NULL, dims = NULL, template = NULL)
}
\arguments{
\item{rvec}{either the column-vectorized data or a list containing it (in element "values")}

\item{dims}{c(nx, ny), the number of x and y grid lines in the full grid}

\item{template}{optional RasterLayer to use as template (for setting crs, resolution etc)

Produces a RasterLayer from a vector or matrix. This essentially a wrapper for various
\code{raster::raster} calls where the input data is in column-vectorized form.

\code{rvec} can either be a matrix (in which case \code{dims} can be omitted) or its vectorized
equivalent, or a list containing elements returned by \code{pkern_fromRaster}. If the list
as an element named "dims", it replaces any argument supplied to \code{dims}. Other elements
of the list (eg 'crs') are passed on to the RasterLayer construction call as needed.

When template is supplied, it supercedes any properties (eg 'crs') supplied in \code{rvec}.}
}
\value{
a RasterLayer containing the data from \code{rvec}
}
\description{
Convert column-vectorized grid to RasterLayer
}
\examples{
require(raster)
r.in = system.file('external/rlogo.grd', package='raster') |> raster(band=1)
pkern.in = pkern_fromRaster(r.in)
pkern_toRaster(pkern.in)
r.in
}
