% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkern_raster.R
\name{pkern_toraster}
\alias{pkern_toraster}
\title{Convert column-vectorized grid to RasterLayer}
\usage{
pkern_toraster(rvec = NA, dims = NULL, template = NULL)
}
\arguments{
\item{rvec}{either the column-vectorized data or a list with elements "data" and "dims"}

\item{dims}{c(nx, ny), the number of x and y grid lines in the full grid}

\item{template}{optional RasterLayer to use as template (for setting crs, resolution etc)

Produces a RasterLayer from a column-vectorized dataset. This is simply a
wrapper for the call \code{raster(matrix(rvec, ncol=dims[1]), template=template)}
with some error checking and the ability to pass both the data vector and its
dimensions together as a list (in "rvec").

If \code{rvec} is a matrix, \code{dims} can be omitted}
}
\value{
a RasterLayer containing the data from "rvec"
}
\description{
Convert column-vectorized grid to RasterLayer
}
\examples{
require(raster)
r.in = system.file('external/rlogo.grd', package='raster') |> raster(band=1)
pkern.in = pkern_fromraster(r.in)
pkern_toraster(pkern.in, template=r.in)
r.in
}
