% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkern_covariance.R
\name{pkern_unpack}
\alias{pkern_unpack}
\title{Vectorize x and y component parameters}
\usage{
pkern_unpack(pars, kp)
}
\arguments{
\item{pars}{list of "x" and "y" kernel parameter lists (each containing vector "kp")}

\item{kp}{vector of combined kernel parameters (or NA)}
}
\value{
numeric vector of parameters, or \code{pars} modified appropriately
}
\description{
when \code{kp} is NULL, the function concatenates the two (x and y) component kernel parameter
sets into a single vector. When this vector is passed back as \code{kp}, the function does the
inverse, copying parameter values from a vector to the corresponding list entry in \code{pars}
and returning them in a list
}
\examples{
pars = pkern_corr(c('mat', 'exp'))
pkern_unpack(pars)
pkern_unpack(pars, 1:3)
pkern_unpack = function(pars, p=NULL)
}
