% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkern_variograms.R
\name{pkern_vario}
\alias{pkern_vario}
\title{Directional sample semivariograms for regular 2D gridded data}
\usage{
pkern_vario(
  dims,
  vec,
  sep = NA,
  simple = TRUE,
  method = "median",
  diagonal = TRUE,
  dmax = NA
)
}
\arguments{
\item{dims}{vector c(nx, ny) of positive integers, the number of grid lines}

\item{vec}{numeric vector of data, in column vectorized order (length \code{prod(dims)})}

\item{sep}{list c(dx, dy) of positive integer vectors, grid line lags to sample}

\item{simple}{logical, if FALSE the function includes point pair indices in return list}

\item{method}{character, one of "mean", "median", "ch", passed to \code{pkern_xvario}}

\item{diagonal}{logical, if TRUE, variogram results are returned also for diagonals}
}
\value{
list of \code{pkern_xvario} output "x", "y" (and possibly "d1", "d2") and sample variance "v"
}
\description{
A wrapper for calls to \code{pkern_xvario} to get the directional sample semivariograms
from regular lattice data along the x and y dimensions, and optionally along the
diagonals (x and y axes after a 45 degree rotation).
}
\details{
Note that \code{sep} is given in terms of grid line indices (integers), not distances
(their product with the x or y resolution). For the diagonals, the grid line separations
are scaled so that \code{sep} is \code{sqrt(2)} for diagonally adjacent grid points.

The function makes the appropriate calls to \code{pkern_xvario} and returns the results
in list elements "x", "y" (and if \code{diagonal=TRUE}, "d1", "d2" the x and y axes after
rotation), along with "v", the estimated (univariate) variance of \code{vec}
}
