% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkern_variograms.R
\name{pkern_vario}
\alias{pkern_vario}
\title{Directional sample semivariograms for regular 2D gridded data}
\usage{
pkern_vario(dims, vec, sep = NA, nmin = 2, simple = TRUE, method = "median")
}
\arguments{
\item{dims}{vector c(nx, ny) of positive integers, the number of grid lines}

\item{vec}{numeric vector of data, in column vectorized order (length \code{prod(dims)})}

\item{sep}{list c(dx, dy) of positive integer vectors, grid line lags to sample}

\item{nmin}{positive integer, the minimum number of point pairs needed to include an estimate}

\item{simple}{logical, if FALSE the function includes point pair indices in return list}

\item{method}{character, one of "mean", "median", "ch", passed to \code{pkern_xvario}}
}
\value{
list with elements "x" and "y" (returned from \code{pkern_xvario}) and "sig"
}
\description{
A wrapper for two calls to \code{pkern_xvario} to get the directional sample semivariograms
along the x and y dimensions, on regular lattice data. Note that \code{sep} is given in terms
of grid line indices (integers), not distances (their product with the x or y resolution)
}
\details{
The function returns its two (x and y component) results in a list along with
"v" the estimated (univariate) variance of \code{vec}
}
