% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkern_variograms.R
\name{pkern_vario}
\alias{pkern_vario}
\title{Directional sample semivariograms for regular 2D gridded data}
\usage{
pkern_vario(
  dims,
  vec,
  sep = NA,
  simple = TRUE,
  method = "median",
  diagonal = TRUE,
  dmax = NA,
  ds = NA
)
}
\arguments{
\item{dims}{vector c(nx, ny) of positive integers, the number of grid lines}

\item{vec}{numeric vector of data, in column vectorized order (length \code{prod(dims)})}

\item{sep}{list c(dx, dy) of positive integer vectors, grid line lags to sample}

\item{simple}{logical, if FALSE the function includes point pair indices in return list}

\item{method}{character, one of "mean", "median", "ch", passed to \code{pkern_xvario}}

\item{diagonal}{logical, if TRUE, semivariogram results are returned also for diagonals}

\item{dmax}{numeric, maximum grid line separation distance (used to filter results)}

\item{ds}{length-2 vector of positive numeric, the spacing distance of x and y grid lines}
}
\value{
list of \code{pkern_xvario} output "x", "y" (and "d1", "d2"); sample variance "v"; and "ds"
}
\description{
A wrapper for calls to \code{pkern_xvario} to get the directional sample semivariograms
from regular lattice data along the x and y dimensions, and optionally along the
diagonals.
}
\details{
The function makes the appropriate calls to \code{pkern_xvario} and returns the results
in list elements "x", "y". If \code{diagonal=TRUE}, the function also returns "d1" and
"d2", the x and y semivariograms after a 45 degree counterclockwise rotation of the
input grid (see \code{pkern_r45}), assuming equal x and y grid line spacings. For unequal
grid line spacings, the true rotation angle is \code{atan(ds[2]/ds[1])}

Note that input \code{sep} is specified in terms of grid line indices (integers), not
distances (their product with the x or y resolution). However, non-integer separation
distances can be specified by setting the resolution of the input grid (dx, dy) in
argument \code{ds}. When \code{ds} is supplied, the output "sep" values are scaled accordingly
and \code{dmax} is interpreted as a distance, rather than a number of grid lines.

When \code{ds} is not supplied, the function assumes that the x grid lines have the same
spacing distance (1) as the y grid lines. In this case, diagonally adjacent
grid points have separation distance \code{sqrt(2)}.
}
