% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkern_variograms.R
\name{pkern_vario_fit}
\alias{pkern_vario_fit}
\title{Fit a theoretical separable covariance model to a separated empirical semivariogram}
\usage{
pkern_vario_fit(
  vario,
  ypars = "gau",
  xpars = ypars,
  psill = NULL,
  nug = NULL,
  add = 0,
  dmax = Inf,
  fit.method = "n/v"
)
}
\arguments{
\item{vario}{list, the return value from a call to \code{pkern_vario}}

\item{ypars}{character or list, recognizable (as \code{pars}) by \code{pkern_corr}}

\item{xpars}{character or list, recognizable (as \code{pars}) by \code{pkern_corr}}

\item{psill}{length-3 numeric vector of bounds (lower, initial, upper) for the partial sill}

\item{nug}{length-3 numeric vector of bounds (lower, initial, upper) for the nugget effect}

\item{add}{positive integer, the number of initial parameter sets tested (see details)}

\item{dmax}{positive numeric, point pairs with separation distances exceeding \code{dmax} are ignored}

\item{fit.method}{character string, either '1', 'n', 'n/d', or 'n/v' (the default, see details)}
}
\value{
TODO
}
\description{
Pipe the results of a \code{pkern_vario} call to this function to estimate model
parameters for \code{xpars} and \code{ypars}, the x and y component kernels, by weighted
least squares using \code{stats::optim} ("L-BFGS-B") for numerical minimization.
}
\details{
\code{xpars} and \code{ypars} may be specified as character strings (the exponential
"exp" is the default), in which case they are assigned suggested initial values
and bounds. If parameter lists (recognized by \code{pkern_corr}), are supplied, the
function uses its "kp" entry for initial values, and if the list contains entries
"lower" and/or "upper", they are used as bounds.

Sill and nugget are estimated simultaneously with the correlation kernel parameters
using the bounds and initial values in \code{psill} and \code{nug}. If these are not
supplied, the function sets (for both) default default bounds of 0 to 4 times the
sample variance.

The nugget effect can be disabled by setting its maximum to 0. Note that for
certain sampling configurations and kernels (particularly the Gaussian), this will
produce ill-conditioned covariance matrices, leading to inaccuracies or errors
in prediction functions like \code{pkern_cmean}.

When \code{add > 0}, the least squares optimization call is repeated for additional
sets of initial values, chosen uniformly at random within their upper and lower
bounds. The best result of these is selected an returned. Try increasing \code{add}
to resolve issues of convergence to incorrect local optima.

\code{fit.method} controls the type of weights assigned to the sum of squares from each
spatial lag: '1' assigns unit weight to all lags (OLS); 'n' weights be sample size;
'n/d' uses sample size divided by squared distance (the default in \code{gstat}); and
'n/v' (the default) uses the robust estimator of Cressie (1993, Ch. 2.6.2), which
divides sample size by squared theoretical semivariance.
}
\examples{
gdim = c(1e2, 2e2)
pars.1d = pkern_corr('gau')
pars.2d = list(y=pars.1d, x=pars.1d, psill=1)
vec = pkern_sim(pars.2d, gdim)

# sample first 10 lags and plot
vario = pkern_vario(gdim, vec, lags=seq(10), nmax=1e3, quiet=TRUE)
pkern_vario_plot(vario, pars.2d)

# fit the separable gaussian kernel
pars.fitted = pkern_vario_fit(vario)
pkern_vario_plot(vario, pars.fitted)
pkern_plot(pars.fitted)

# fit a different separable kernel
pars.fitted2 = pkern_vario_fit(vario, ypars='sph', xpars='mat')
pkern_vario_plot(vario, pars.fitted2)
pkern_plot(pars.fitted2)

}
