% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkern_variograms.R
\name{pkern_vario_fit}
\alias{pkern_vario_fit}
\title{Fit a theoretical separable covariance model to a separated empirical semivariogram}
\usage{
pkern_vario_fit(
  vario,
  xpars = "exp",
  ypars = xpars,
  v = NA,
  nug = NA,
  nmin = 2,
  ninitial = 1
)
}
\arguments{
\item{vario}{list, the return value from a call to \code{pkern_vario}}

\item{xpars}{character or list, recognizable (as \code{pars}) by \code{pkern_corr}}

\item{ypars}{character or list, recognizable (as \code{pars}) by \code{pkern_corr}}

\item{v}{length-3 numeric vector, c('min', 'ini', 'max'), bounds and initial value for variance}

\item{nug}{length-3 numeric vector, c('min', 'ini', 'max'), bounds and initial value for nugget}

\item{nmin}{positive integer, lags with fewer than \code{nmin} samples are omitted in estimation}

\item{ninitial}{positive integer, the number of initial parameter sets tested (see details)}
}
\value{

}
\description{
Pipe the results of a \code{pkern_vario} call to this function to estimate model
parameters for \code{xpars} and \code{ypars}, the x and y component kernels, by weighted
least squares (Cressie, 2015), where \code{stats::optim} ("L-BFGS-B") with default
settings is used to solve the minimization problem.
}
\details{
\code{xpars} and \code{ypars} may be specified as character strings (the exponential
"exp" is the default), in which case they are assigned suggested initial values
and bounds. If parameter lists (recognized by \code{pkern_corr}), are supplied, the
function uses its "kp" entry for initial values, and if the list contains entries
"lower" and/or "upper", they are used as bounds.

The pointwise variance is estimated simultaneously with the correlation kernel
parameters using the bounds and initial values in \code{v} and \code{nug}. If these are not
supplied, the function sets (for both) default default bounds of 0 to 4 times the
sample variance.

The nugget effect can be disabled by setting its maximum to 0. Note that for
certain sampling configurations and kernels (particularly the Gaussian), this will
produce ill-conditioned covariance matrices, leading to inaccuracies or errors
in prediction functions like \code{pkern_cmean}.

When \code{ninitial > 1}, the least squares optimization call is repeated for additional
sets of initial values, chosen uniformly at random within their upper and lower
bounds. Try increasing this to resolve issues of convergence to incorrect local optima.
}
