% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkern_variograms.R
\name{pkern_vario_plot}
\alias{pkern_vario_plot}
\title{Plot 1-dimensional empirical semivariograms for a separable model}
\usage{
pkern_vario_plot(vario, pars = NULL, plotpars = NULL)
}
\arguments{
\item{vario}{list of semivariance data, the return value of \code{pkern_vario} or \code{pkern_xvario}}

\item{pars}{list of kernel parameters "x" and/or "y", variance "v", and optionally nugget "nug"}

\item{plotpars}{list, plotting options (see details)}
}
\value{
prints a plot and returns nothing
}
\description{
Plots the output of \code{pkern_vario} or \code{pkern_xvario}, optionally including the theoretical
semivariogram values for a model specified in \code{pars} (eg the return value of \code{pkern_vario_fit}).
Note that the plot shows semivariance (gamma), whereas the "vg" elements returned by
\code{pkern_vario} are variogram values (2*gamma).
}
\details{
\code{pars} should be a list containing named elements: "y" and "x", lists of y and x kernel
parameters in form recognized by \code{pkern_corr}; "psill", the partial sill and, optionally,
"nug", the nugget variance.

\code{plotpars} applies to all plots. It should be a list containing any of the following elements:

"title" character or vector of two, title(s) for the plot(s)
"shade" logical, indicating to shade points according to the number of samples
"dmax" positive numeric, upper limit for x in scatterplot
"vglim" length-2 numeric vector, the y limits for the scatterplot
}
\examples{
gdim = c(100, 50)
pars.1d = pkern_corr('gau')
pars.2d = list(y=pars.1d, x=pars.1d, psill=1)
vec = pkern_sim(pars.2d, gdim)

# sample first 10 lags and plot results
vario = pkern_vario(gdim, vec, lags=seq(10), quiet=TRUE)
pkern_vario_plot(vario, pars.2d)

# plot componentwise
pkern_vario_plot(vario[['x']], c(pars.1d, list(psill=1)))
pkern_vario_plot(vario[['y']], c(pars.1d, list(psill=1)))

}
