% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkern_indexing.R
\name{pkern_vec2mat}
\alias{pkern_vec2mat}
\title{Inverse column-vectorization indices}
\usage{
pkern_vec2mat(idx, m)
}
\arguments{
\item{idx}{a vector of positive integers}

\item{m}{number of rows in the matrix}
}
\value{
a two column matrix of integers (row and column numbers) with length(idx) rows
}
\description{
Column vectorization (as in base::as.vector) builds a length(mn) vector by stacking
the columns of an m X n matrix. The (i,j)th element of the matrix is mapped to the
\code{(i + m*(j-1))}th element of the vector. This function is the inverse of that mapping.
It finds the (i,j) associated with a given element after column-vectorization
}
\examples{

m = 5
n = 6
matrix.indexing = matrix(1:prod(m, n), m)
as.vector(matrix.indexing)
pkern_vec2mat(2, m) # finds the matrix indices for the second component
pkern_vec2mat(6:10, m) # the function is vectorized
}
