% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkern_variograms.R
\name{pkern_xvario}
\alias{pkern_xvario}
\title{Sample semivariogram for regular 2D gridded data along x direction}
\usage{
pkern_xvario(dims, vec, sep = NA, simple = TRUE, method = "median")
}
\arguments{
\item{dims}{vector c(nx, ny) of positive integers, the number of grid lines in full grid}

\item{vec}{numeric vector of data, in column vectorized order (length \code{prod(dims)})}

\item{sep}{positive integer vector, the grid line lags to sample}

\item{simple}{logical, if FALSE the function returns a list of point pair indices}

\item{method}{character, one of "mean", "median", "ch"}
}
\value{
a list with named elements "sep", "vario", and "pairs"
}
\description{
The function identifies and computes the semivariance (in \code{vec}) of point pairs on a
regular grid of dimensions \code{dims} at the supplied x grid line lags, \code{sep}. Elements of
\code{set} must be a subset of \code{1:(dims[1]-1)} (the default setting).
}
\details{
Results are returned in a named list containing the lags ("sep"), the number of point
pairs ("n") and sample semivariances ("sv") at each lag, and optionally (if \code{simple}
is FALSE) a list of matrices containing the indices of point pairs sampled and
their absolute difference ("pairs").

method "mean" is the classical method of Matheron (1962), "median" and "ch" are the
median and robust methods described in Cressie and Hawkins (1980).

references:

Cressie and Hawkins (1980): https://doi.org/10.1007/BF01035243
}
\examples{


}
